### 2021-06-29 #41 @jonmatthis
Help request - Prepare a Python repository for submission
## What is this?

To make the review process easier and to help you maintain your package, pyOpenSci requires testing, continuous integration, and full documentation for packages submitted for review (read more [here][PackagingGuide]. We want to make this as easy as possible for newcomers, so if you'd like some help in prepping your package you can submit this request and we'll assign someone to help you out!

Before submitting a request, check out our [Packaging Guide][PackagingGuide]. The answers to your questions might be there already.

## Package Info
Submitting Author: Jon Matthis (@jonmatthis)
Repository Link (if existing):  https://github.com/jonmatthis/freemocap

---

```
The FreeMoCap Project: A free-and-open-source, hardware-and-software-agnostic, research-grade, minimal-cost motion capture system and platform for decentralized scientific research, education, and training

```

- If you selected Unsure/Other, briefly explain:

		This package is intended to serve as the backbone/framework for a large, long term project with a variety of scientific and educational goals that cut across many of the categories listed here. I'd be happy to just pick one, if that's better for your system though!
## How Can We Help?

- Indicate how we can help you prepare your package (select all that apply):
	- [X] Creating a test suite
	- [X] Enabling Continuous Integration
	- [ ] Adding a specific feature
	- [X] Improving Documentation
	- [X] General Assistance
	- [X] Other

- Explain briefly below if there's something specific you'd like help with:

	I have 10+ years experience working in Matlab, but this is my first Python project. Neither I nor anyone on my team (of mostly undergrads) have any significant software development XP (beyond the classical "Dropbox folder full of Matlab .m files per publication" model of academic software development 😅). Any help would be greatly appreciated!

I followed the instructions on this page to get this repo on to PyPi shortly before I discovered pyOpenSci - https://github.com/NSLS-II/scientific-python-cookiecutter

I'm so happy you exist! I've been looking for something this (+ JoSS) for a long time!!
