+++ freemocap_blender_addon/freemocap_blender_addon/old_loadOpenMocapSession.py
@@ -1,0 +1,184 @@
+import bpy
+import numpy as np
+from mathutils import Matrix, Vector, Euler
+import pathlib 
+
+class OMC_OT_loadOpenMoCapSession(bpy.types.Operator): #setting the type as "OMC" for "OpenMoCap"
+    """ Load data from an OpenMoCap(OMC) recording session into Blender """
+    bl_idname = "omc.load_session_data"
+    bl_label = "Load Session Data"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    charucoMar_size: bpy.props.FloatProperty(
+        name = 'Charuco Marker Size',
+        description="Radius of the Charuco Marker Spheres",
+        default = .2,
+        min = 0, soft_max = 1,
+    )
+    
+    dlcMar_size: bpy.props.FloatProperty(
+        name = 'DLC Marker Size',
+        description="Radius of the DLC Marker Spheres",
+        default = .3,
+        min = 0, soft_max = 1,
+    )
+
+    bodyMar_size: bpy.props.FloatProperty(
+        name = 'Skeleton Body Marker Size',
+        description="Radius of the Skeleton Body Marker Spheres",
+        default = .2,
+        min = 0, soft_max = 1,
+    )
+
+    faceMar_size: bpy.props.FloatProperty(
+        name = 'Skeleton Face Marker Size',
+        description="Radius of the Skeleton Face Marker Spheres",
+        default = .15,
+        min = 0, soft_max = 1,
+    )
+
+    handMar_size: bpy.props.FloatProperty(
+        name = 'Skeleton Hand Marker Size',
+        description="Radius of the Skeleton Hand Marker Spheres",
+        default = .2,
+        min = 0, soft_max = 1,
+    )
+
+    meshType: bpy.props.StringProperty(
+        name = 'Marker Mesh Type',
+        description = 'Type of mesh to draw at marker locations (just `sphere` and `monkey` for now)',
+        default = 'uv_sphere',
+    )
+
+    def execute(self, context):
+        """Load in data from an OpenMoCap recording session and place all 3D points as spheres in the 3D Viewport"""
+        C = bpy.context
+        D = bpy.data
+
+        # #%% Nuke it all 
+        # bpy.ops.object.select_all(action='SELECT') #select all
+        # bpy.ops.object.delete(use_global=False) #delete everything 
+
+        # try:
+        #     bpy.ops.outliner.select_all(action='SELECT')
+        #     bpy.ops.outliner.delete(hierarchy=True)
+        # except:
+        #     pass
+
+        #%% _______________________________________________________________________
+        #  Load in data
+
+        loadedData  = np.load(r"C:\Users\jonma\Dropbox\GitKrakenRepos\OpenMoCap\Data\test6_01_21a\outData\saveData.npz")
+        skel_fr_mar_dim = loadedData['skel_fr_mar_dim'] #OpenPose skeleton data
+        dlcPts_fr_mar_dim = loadedData['dlcPts_fr_mar_dim']
+        charucoCorners_pt_XYZ = loadedData['charucoCornersXYZ']
+
+        startFr = 0
+        numFrames = skel_fr_mar_dim.shape[0]-1
+
+        #names of markers 
+        skel_markerID = ["Nose", "Neck", "RShoulder", "RElbow", "RWrist", "LShoulder",
+        "LElbow", "LWrist", "MidHip", "RHip", "RKnee", "RAnkle", "LHip", "LKnee",
+        "LAnkle", "REye", "LEye", "REar", "LEar", "LBigToe", "LSmallToe", "LHeel",
+        "RBigToe", "RSmallToe", "RHeel"]
+
+        #%% create spheres at marker locations
+
[…]
