### 2021-06-30 8d660aa
dumped pip freeze into requirements.txt
This is too restrictive, but maybe worth a shot for now
