### 2021-06-30 07ea1df
Stable Stage 1 and 2 changes
Session class overhaul
Stage 2 time syncing assertions
