+++ freemocap/session.py
@@ -1,8 +1,20 @@
-class Session: #session like "recording session"
+from pathlib import Path
+from ruamel.yaml import YAML
+
+from freemocap import recordingconfig
+
+class Session: #self like "recording self"
     def __init__(self):
+        
+        self.numCams = None
+        self.numFrames = None
+        self.numTrackedPoints = None 
+
+        self.basePath = Path.cwd()
+        
         self.sessionID = '' #The sessionID tag will be used to generate files names and whatnot
         self.sessionPath = '' #The folder where the to-be-processed videos live (in a folder called "synced Vids")
-        self.numCams = ''#The number of cameras used in this recording session
+        #self.numCams = ''#The number of cameras used in this recording self
         self.openPoseDataPath   = ''#Where the open pose data lives
         self.dlcDataPath    = ''#Where the DLC data lives
         self.openPoseExePath = r'C:\openpose'
@@ -18,3 +30,160 @@ class Session: #session like "recording session"
 
         self.mediaPipe_imgPathList = None
         self.openPose_imgPathList = None
+
+    def start_session(self,paramDict,rotDict):
+        #start a session, create all the file paths necessary, and create a session dictionary to save settings
+        recordPath = self.basePath/'Data'
+        recordPath.mkdir(exist_ok= True)
+
+        self.sessionPath = recordPath/self.sessionID
+        self.sessionPath.mkdir(exist_ok=True)
+
+        #create all the session filepaths and settings - create and load them into a dictionary
+        self.pathList = self.create_session_paths()
+        self.session_settings = self.create_session_dictionary()
+        self.create_session_txt(paramDict,rotDict)
+
+
+
+    def create_session_paths(self):
+        #creates Path objects to each folder created in a recording self and adds them all to the pathList variable
+        #folders for these paths are created at the start of each stage where they are needed 
+
+        pathList = []
+
+        self.rawVidPath = self.sessionPath/'RawVideos'
+        pathList.append(self.rawVidPath)
+
+        self.syncedVidPath = self.sessionPath/'SyncedVideos'
+        pathList.append(self.syncedVidPath)
+
+        self.calVideos = self.sessionPath/'CalVideos'
+        pathList.append(self.calVideos)
+
+        self.dataArrays = self.sessionPath/'DataArrays'
+        pathList.append(self.dataArrays)
+
+        self.DLCData = self.sessionPath/'DLCData'
+        pathList.append(self.DLCData)
+
+        self.openPoseDataPath = self.sessionPath/'OpenPoseData'
+        pathList.append(self.openPoseDataPath)
+
+        self.mediaPipeDataPath = self.sessionPath/'MediaPipeData'
+        pathList.append(self.mediaPipeDataPath)
+
+        self.imOutPath = self.sessionPath/'imOut'
+        pathList.append(self.imOutPath)
+        
+        return pathList
+
+    def create_session_dictionary(self):
+        #creates a dictionary of settings that should be saved into a YAML throughout each self
+
+        recording_parameters = {'numCams': self.numCams, 'numFrames': self.numFrames, 'numTrackedPoints': self.numTrackedPoints}
+
+        #create a dictionary with the folder names (taken from the recordingconfig.py file) and the folder paths from pathList
+        #NOTE -6/30/21 - as of right now, the order of folder names in the recordingconfig.folder_setup list needs to match the 
+        #                the order of folder paths in the pathList
+        session_paths = {}
+
+        for foldername, folderpath in zip(recordingconfig.folder_setup,self.pathList):
+            session_paths[foldername] = folderpath
+
+        session_settings_dictionary = {'recording_parameters':recording_parameters, 'session_paths':session_paths}
+        
[…]
