### 2021-06-30 cfd86f7
Stable Stage 1 and 2 changes
Session class overhaul
Stage 2 time syncing assertions
