+++ freemocap/runcams.py
@@ -10,7 +10,9 @@ from tkinter import Tk
 
 
 def RecordCams(session,camInputs,parameterDictionary,rotationInputs):
-   
+    #Create RawVideos folder
+    session.rawVidPath.mkdir(exist_ok = True)
+
     #%% Setting up recordings
     beginTime = time.time()
     numCams = len(camInputs) #number of cameras 
@@ -58,15 +60,17 @@ def RecordCams(session,camInputs,parameterDictionary,rotationInputs):
     timeStampData.to_csv(csvPath) #turn dataframe into a CSV
 
     session.numCams = numCams
+    session.session_settings['recording_parameters'].update({'numCams':session.numCams})
     session.timeStampData = timeStampData
     session.camIDs = camIDs
     session.numCamRange = numCamRange
     session.vidNames = vidNames
 
 def SyncCams(session, timeStampData,numCamRange,vidNames,camIDs):
+    session.syncedVidPath.mkdir(exist_ok = True)
 
     #start the timesync process
-    frameTable,timeTable,frameRate,resultsTable,plots = timesync.TimeSync(timeStampData,numCamRange,camIDs) 
+    frameTable,timeTable,frameRate,resultsTable,plots = timesync.TimeSync(session,timeStampData,numCamRange,camIDs) 
     
     #this message shows you your percentages and asks if you would like to continue or not. shuts down the program if no
     root = Tk()
@@ -77,5 +81,6 @@ def SyncCams(session, timeStampData,numCamRange,vidNames,camIDs):
         print()
         print('Starting editing')
         videotrim.VideoTrim(session,vidNames,frameTable,session.parameterDictionary,session.rotationInputs,numCamRange)
+        session.session_settings['recording_parameters'].update({'numFrames':session.numFrames})
         #videotrim.createCalibrationVideos(session,60,parameterDictionary)
         print('all done')
