+++ freemocap/recordingconfig.py
@@ -4,138 +4,141 @@ Created on Thu Mar 11 11:54:12 2021
 
 @author: Rontc
 """
-from ruamel.yaml import YAML
-from pathlib import Path
 
-def createSession(session,filepath, existing_data = False):
-    #create file paths for the session, raw and synced videos
+folder_setup = ['RawVideos','SyncedVideos','CalVideos','DataArrays','DLCData','OpenPoseData','MediaPipeData','imOut']
+
+
+
+
+# def createSession(session,filepath, existing_data = False):
+#     #create file paths for the session, raw and synced videos
     
-    pathList = []
+#     pathList = []
    
-    recordPath = filepath/'Data'
-    recordPath.mkdir(exist_ok= True)
+#     recordPath = filepath/'Data'
+#     recordPath.mkdir(exist_ok= True)
     
-    session.sessionPath = recordPath/session.sessionID
-    #if existing_data:
-        #session.sessionPath.mkdir(exist_ok = True)
-    #else:
-    session.sessionPath.mkdir(exist_ok= True)
-    pathList.append(session.sessionPath)
+#     session.sessionPath = recordPath/session.sessionID
+#     #if existing_data:
+#         #session.sessionPath.mkdir(exist_ok = True)
+#     #else:
+#     session.sessionPath.mkdir(exist_ok= True)
+#     pathList.append(session.sessionPath)
     
-    session.rawVidPath = session.sessionPath/'RawVideos'
-    session.rawVidPath.mkdir(exist_ok = True)
-    pathList.append(session.rawVidPath)
+#     session.rawVidPath = session.sessionPath/'RawVideos'
+#     session.rawVidPath.mkdir(exist_ok = True)
+#     pathList.append(session.rawVidPath)
 
-    session.syncedVidPath = session.sessionPath/'SyncedVideos'   
-    session.syncedVidPath.mkdir(exist_ok = True)
-    pathList.append(session.syncedVidPath)
+#     session.syncedVidPath = session.sessionPath/'SyncedVideos'   
+#     session.syncedVidPath.mkdir(exist_ok = True)
+#     pathList.append(session.syncedVidPath)
 
-    session.calVidPath = session.sessionPath/'CalVideos'
-    session.calVidPath.mkdir(exist_ok = True)
-    pathList.append(session.calVidPath)
+#     session.calVidPath = session.sessionPath/'CalVideos'
+#     session.calVidPath.mkdir(exist_ok = True)
+#     pathList.append(session.calVidPath)
 
-    session.openPoseDataPath = session.sessionPath/'OpenPoseData'
-    session.openPoseDataPath.mkdir(exist_ok = True)
-    pathList.append(session.openPoseDataPath)
+#     session.openPoseDataPath = session.sessionPath/'OpenPoseData'
+#     session.openPoseDataPath.mkdir(exist_ok = True)
+#     pathList.append(session.openPoseDataPath)
 
-    session.mediaPipeDataPath = session.sessionPath/'MediaPipeData'
-    session.mediaPipeDataPath.mkdir(exist_ok = True)
-    pathList.append(session.mediaPipeDataPath)
+#     session.mediaPipeDataPath = session.sessionPath/'MediaPipeData'
+#     session.mediaPipeDataPath.mkdir(exist_ok = True)
+#     pathList.append(session.mediaPipeDataPath)
     
-    session.dlcDataPath  =  session.sessionPath / 'DLCdata'
-    session.dlcDataPath.mkdir(exist_ok = True, parents = True)
-    pathList.append(session.dlcDataPath)
+#     session.dlcDataPath  =  session.sessionPath / 'DLCdata'
+#     session.dlcDataPath.mkdir(exist_ok = True, parents = True)
+#     pathList.append(session.dlcDataPath)
     
[…]
