### 2021-06-30 99c684b
Split up webcamGUI initialization, Stage 3 fixes
Pipeline now works through Stage 3
