### 2021-06-30 2d876e4
Split up webcamGUI initialization, Stage 3 fixes
Pipeline now works through Stage 3
