+++ parameters.yaml
@@ -20,11 +20,10 @@ default:
 
 saved:
   rotations:
-    0: 0
     1: 0
     2: 0
   parameters:
-    exposure: -5
+    exposure: -6
     resWidth: 640
     resHeight: 480
     framerate: 30
