+++ freemocap/parameters.yaml
@@ -1,0 +1,27 @@
+#default parameters
+
+
+
+default:
+  rotations: {}
+  parameters:
+    exposure: -5
+    resWidth: 640
+    resHeight: 480
+    framerate: 25
+    codec: DIVX
+
+#saved parameters    
+
+
+
+
+saved:
+  rotations:
+    0: 0
+  parameters:
+    exposure: -6
+    resWidth: 640
+    resHeight: 480
+    framerate: 30
+    codec: H264
