### 2021-07-01 4076fa9
Getting openpose working
Pipeline works from most stages now (except 2) with openpose
