### 2021-07-01 a9b8e33
Getting openpose working
Pipeline works from most stages now (except 2) with openpose
