+++ freemocap/parameters.yaml
@@ -1,4 +1,4 @@
-#default parameters
+#default parameters
 
 
 
@@ -11,14 +11,15 @@ default:
     framerate: 25
     codec: DIVX
 
-#saved parameters    
+#saved parameters    
 
 
 
 
 saved:
   rotations:
-    0: 0
+    1: 0
+    2: 0
   parameters:
     exposure: -6
     resWidth: 640
