### 2021-07-01 353d8a0
openpose img/json path saving fixes
updated how openpose img/json paths are saved to be more friendly with our new config saving methods
