+++ runme_fmc_staged.py
@@ -1,167 +1,0 @@
-from freemocap import createvideo, initialization, runcams, calibrate, fmc_mediapipe, fmc_openpose, fmc_deeplabcut, reconstruct3D, playskeleton, session
-from freemocap.fmc_pyqtgraph import PlayerDockedWindow
-
-from pathlib import Path
-import os
-
-from aniposelib.boards import CharucoBoard
-
-import numpy as np
-
-sesh = session.Session()
-
-
-sesh.useOpenPose= True
-sesh.useMediaPipe = False
-sesh.useDLC=True
-
-if sesh.useDLC: 
-    # import deeplabcut as dlc
-    sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
-
-
-
-# %% Inputs to edit
-stage = 1 #set your starting stage here (stage = 1 will run the pipeline from webcams)
-sesh.debug = False
-
-sesh.sessionID = 'sesh_21-05-31_111833' #fill in if you are running from Stage 2 onwards
-if not sesh.sessionID:
-    dataFolder = Path.cwd()/'Data'
-    subfolders = [ f.path for f in os.scandir(dataFolder) if f.is_dir() ] #copy-pasta from who knows where
-    sesh.sessionID = Path(subfolders[-1]).stem #grab the name of the last folder in the list of subfolders
-
-board = CharucoBoard(7, 5,
-                     #square_length=1, # here, in mm but any unit works (JSM NOTE - just using '1' so resulting units will be in 'charuco squarelenghts`)
-                     #marker_length=.8,
-                     square_length = 121,
-                     marker_length = 96,
-                     marker_bits=4, dict_size=250)
-
-
-
-#sesh.input_stage = stage
-
-
-
-# %% Initialization
-
-initialization.initialize(sesh,stage,board)
-
-# %% Stage One
-if stage <= 1:
-    print()
-    print('Starting Video Recordings')
-    runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
-else:
-    print('Skipping Video Recording')
-
-
-# %% Stage Two
-if stage <= 2:
-    print()
-    print('Starting Video Syncing')
-    runcams.SyncCams(sesh, sesh.timeStampData,sesh.numCamRange,sesh.vidNames,sesh.camIDs)
-else:
-    print('Skipping Video Syncing')
-
-# %% Stage Three
-if stage <= 3:
-    print()
-    print('Starting Calibration')
-    sesh.cgroup, sesh.mean_charuco_fr_mar_dim = calibrate.CalibrateCaptureVolume(sesh,board)
-else:
-    print('Skipping Calibration')
-
-# %% Stage Four
-if stage <= 4:
-    print('Starting Track Image Points')
-    if sesh.useMediaPipe:
-        fmc_mediapipe.runMediaPipe(sesh)
-        sesh.mediaPipeData_nCams_nFrames_nImgPts_XYC = fmc_mediapipe.parseMediaPipe(sesh)
-        sesh.mediaPipeSkel_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.mediaPipeData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=.1)
-        np.save(sesh.dataArrayPath/'mediaPipeSkel_3d.npy', sesh.mediaPipeSkel_fr_mar_dim) #save data to npy
-
-    if sesh.useOpenPose:
-        fmc_openpose.runOpenPose(sesh, dummyRun=True)
[…]
