+++ requirements.txt
@@ -1,112 +1,20 @@
-# List required packages in this file, one per line.
-numpyabsl-py==0.12.0
-aniposelib==0.4.1
-astunparse==1.6.3
-attrs==21.2.0
-backcall==0.2.0
-cached-property==1.5.2
-cachetools==4.2.2
-certifi==2021.5.30
-chardet==4.0.0
-click==8.0.1
-colorama==0.4.4
-commonmark==0.9.1
-cycler==0.10.0
-Cython==0.29.23
-decorator==4.4.2
 -e git+https://github.com/jonmatthis/DeepLabCut.git@74d3234efd8b7eca5dc317aee3c4f26dbf29a5b5#egg=deeplabcut
-filterpy==1.4.5
-flatbuffers==1.12
-gast==0.4.0
-google-auth==1.30.2
-google-auth-oauthlib==0.4.4
-google-pasta==0.2.0
-grpcio==1.34.1
-h5py==3.1.0
-idna==2.10
-imageio==2.9.0
-imageio-ffmpeg==0.4.4
-imgaug==0.4.0
-importlib-metadata==4.5.0
+aniposelib==0.4.1
+h5py==3.3.0
 imutils==0.5.4
-intel-openmp==2021.2.0
-ipython==7.24.1
-ipython-genutils==0.2.0
-jedi==0.18.0
-joblib==1.0.1
-keras-nightly==2.5.0.dev2021032900
-Keras-Preprocessing==1.1.2
-kiwisolver==1.3.1
-llvmlite==0.33.0
-Markdown==3.3.4
-matplotlib==3.1.3
-matplotlib-inline==0.1.2
-mediapipe==0.8.5
-moviepy==1.0.1
-msgpack==1.0.2
-msgpack-numpy==0.4.7.1
-networkx==2.5.1
-numba==0.50.1
-numexpr==2.7.3
+matplotlib
+mediapipe==0.8.6
 numpy==1.17.3
-oauthlib==3.1.1
-opencv-contrib-python==3.4.14.51
-opt-einsum==3.3.0
-pandas==1.2.4
-parso==0.8.2
-patsy==0.5.1
-pickleshare==0.7.5
-Pillow==8.2.0
-proglog==0.1.9
-prompt-toolkit==3.0.18
-protobuf==3.17.3
-psutil==5.8.0
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-Pygments==2.9.0
-PyOpenGL==3.1.5
-pyparsing==2.4.7
-PyQt5==5.15.4
-PyQt5-Qt5==5.15.2
-PyQt5-sip==12.9.0
+opencv-contrib-python==3.4.14.51 
+pandas==1.2.5
 pyqtgraph==0.12.1
-python-dateutil==2.8.1
-pytz==2021.1
-PyWavelets==1.1.1
 PyYAML==5.4.1
-pyzmq==22.1.0
-requests==2.25.1
-requests-oauthlib==1.3.0
-rich==10.2.2
-rsa==4.7.2
-ruamel.yaml==0.17.8
-ruamel.yaml.clib==0.2.2
-scikit-image==0.18.1
-scikit-learn==0.24.2
-scipy==1.6.3
-Shapely==1.7.1
-six==1.15.0
-statsmodels==0.12.2
-tables==3.6.1
-tabulate==0.8.9
-tensorboard==2.5.0
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.0
-tensorflow-estimator==2.5.0
-tensorflow-gpu==2.5.0
-tensorpack==0.9.8
-termcolor==1.1.0
-tf-slim==1.1.0
[…]
