+++ parameters.yaml
@@ -21,9 +21,6 @@ default:
 saved:
   rotations:
     0: 90
-    1: 270
-    2: 90
-    3: 270
   parameters:
     exposure: -8
     resWidth: 640
