+++ freemocap/webcam/videotrim.py
@@ -16,17 +16,6 @@ def VideoTrim(session, vidList, ft, parameterDictionary, rotationState, numCamRa
     trueHeight = resHeight
     codec = parameterDictionary.get("codec")
 
-<<<<<<< HEAD
-    for vid, cam, camNum in zip(
-        vidList, camList, numCamRange
-    ):  # iterate in parallel through camera identifiers and matching videos
-        print("Editing " + cam + " from " + vid)
-        cap = cv2.VideoCapture(
-            str(session.rawVidPath / vid)
-        )  # initialize OpenCV capture
-        frameTable = ft[cam]  # grab the frames needed for that camera
-        success, image = cap.read()  # start reading frames
-=======
     postTrimmingTotalNumFrames = []
 
     for vid,cam,camNum in zip(vidList,camList,numCamRange): #iterate in parallel through camera identifiers and matching videos
@@ -34,7 +23,6 @@ def VideoTrim(session, vidList, ft, parameterDictionary, rotationState, numCamRa
         cap = cv2.VideoCapture(str(session.rawVidPath/vid)) #initialize OpenCV capture
         frameTable = ft[cam] #grab the frames needed for that camera
         success, image = cap.read() #start reading frames
->>>>>>> merge-pipy-aaron
         fourcc = cv2.VideoWriter_fourcc(*codec)
         saveName = session.sessionID + "_synced_" + cam + ".mp4"
 
@@ -72,13 +60,6 @@ def VideoTrim(session, vidList, ft, parameterDictionary, rotationState, numCamRa
         resHeight = trueHeight
         cap.release()
         out.release()
-<<<<<<< HEAD
-        print("Saved " + saveSyncedVidPath)
-        print()
-
-
-def createCalibrationVideos(session, calVideoFrameLength, parameterDictionary):
-=======
         frame_length_cap = cv2.VideoCapture(saveSyncedVidPath)
         thisCamNumFrame = int(frame_length_cap.get(cv2.CAP_PROP_FRAME_COUNT)) 
         postTrimmingTotalNumFrames.append(thisCamNumFrame)
@@ -91,7 +72,6 @@ def createCalibrationVideos(session, calVideoFrameLength, parameterDictionary):
 
 
 def createCalibrationVideos(session,calVideoFrameLength,parameterDictionary):
->>>>>>> merge-pipy-aaron
     vidList = os.listdir(session.syncedVidPath)
     framelist = list(range(calVideoFrameLength))
     codec = parameterDictionary.get("codec")
