### 2021-07-02 afd82c0
prepping for v0.0.2
