+++ AUTHORS.rst
@@ -2,12 +2,14 @@
 Credits
 =======
 
-Maintainer
+Maintainers
 ----------
 
 * Jonathan Samir Matthis <jonmatthis@gmail.com>
+Aaron Charian @aaroncherian
 
 Contributors
 ------------
+
 
 None yet. Why not be the first? See: CONTRIBUTING.rst
