### 2021-07-03 5830743
Create codeql-analysis.yml
