### 2021-07-03 82029dd
fixed session initialization
