+++ requirements.txt
@@ -1,20 +1,19 @@
--e git+https://github.com/jonmatthis/DeepLabCut.git@74d3234efd8b7eca5dc317aee3c4f26dbf29a5b5#egg=deeplabcut
-aniposelib==0.4.1
-h5py==3.3.0
-imutils==0.5.4
-matplotlib
-mediapipe==0.8.6
-numpy==1.17.3
-opencv-contrib-python==3.4.14.51 
-pandas==1.2.5
-pyqtgraph==0.12.1
-PyYAML==5.4.1
-rich==10.4.0
-ruamel.base==1.0.0
-scipy==1.7.0
-toml==0.10.2
-tqdm==4.61.1
-tensorflow-gpu
-pyqt5
-pyopengl
-tf_slim
+aniposelib~=0.4.1
+h5py>=3.1.0
+imutils~=0.5.4
+matplotlib~=3.1.3
+mediapipe~=0.8.6
+opencv-contrib-python~=3.4.14.51 
+pandas~=1.2.5
+pyqtgraph~=0.12.1
+rich~=10.4.0
+scipy~=1.7.0
+toml~=0.10.2
+tqdm~=4.61.1
+tensorflow-gpu~=2.0
+pyqt5~=5.15.4
+pyopengl~=3.1.5
+tf_slim~=1.1.0
+deeplabcut[gui]~=2.2rc3
+ruamel.yaml>=0.15.0
+typed-ast<1.5,>=1.4.0
