### 2021-07-04 f81d3cb
add _sample_data_instructions.txt to Manifest
hopefully that means it will include that file, along with the Data/ folder?
