### 2021-07-04 dfab2ef
mkdir for 'FreeMoCap_Data' folder on first run
Also change name of ' Data'  folder to 'FreeMoCap_Data'
