+++ freemocap/parameters.yaml
@@ -1,4 +1,4 @@
-#default parameters
+#default parameters
 
 
 
@@ -12,7 +12,7 @@ default:
     framerate: 25
     codec: DIVX
 
-#saved parameters    
+#saved parameters    
 
 
 
@@ -22,6 +22,8 @@ saved:
   rotations:
     1: 0
     2: 0
+    3: 0
+    4: 0
   parameters:
     exposure: -6
     resWidth: 640
