### 2021-07-04 669abd6
mkdir for 'FreeMoCap_Data' folder on first run
Also change name of ' Data'  folder to 'FreeMoCap_Data'
