### 2021-07-05 d2d8167
pip install works!
run:
>python
>>import freemocap as fmc
>>fmc.RunMe()
