### 2021-07-05 a06177b
pip install works!
run:
>python
>>import freemocap as fmc
>>fmc.RunMe()
