+++ requirements.txt
@@ -10,10 +10,10 @@ rich~=10.4.0
 scipy~=1.7.0
 toml~=0.10.2
 tqdm~=4.61.1
-tensorflow-gpu~=2.0
-pyqt5~=5.15.4
-pyopengl~=3.1.5
-tf_slim~=1.1.0
-deeplabcut[gui]~=2.2rc3
+tensorflow-gpu>=2.0
+pyqt5>=5.15.4
+pyopengl>=3.1.5
+tf_slim>=1.1.0
+deeplabcut[gui]>=2.2rc3
 ruamel.yaml>=0.15.0
 typed-ast<1.5,>=1.4.0
