### 2021-07-05 0c06bda
tweaks during convo with Aaron
