+++ runme_FreeMoCap.py
@@ -59,6 +59,10 @@ if not sesh.sessionID:
         dataFolder.mkdir()
         sampleDataFolder = dataFolder / '_sample_data_folder'
         sampleDataFolder.mkdir()
+        # prompt user if they want to download sample data
+        # if yes, download sample session folder from public URL (i.e. figshare )
+        # (also same for relevant DLC folder, which we should remove from the .git)
+
 
 board = CharucoBoard(
     7,
