+++ freemocap/parameters.yaml
@@ -20,10 +20,10 @@ default:
 
 saved:
   rotations:
-    1: 90
-    2: 90
-    3: 90
-    4: 90
+    0: 0
+    1: 0
+    2: 0
+    3: 0
   parameters:
     exposure: -8
     resWidth: 640
