### 2021-07-05 06c0cf6
merge bug fixes
Fixed errors with outdated lines of code being copied over during a merge
