### 2021-07-06 75a3291
Parameter yaml creation
Deleted extra files, FMC now generates a parameters yaml if none is found
