### 2021-07-06 4c84e1b
Parameter yaml creation
Deleted extra files, FMC now generates a parameters yaml if none is found
