+++ freemocap/recordingconfig.py
@@ -6,77 +6,12 @@ Created on Thu Mar 11 11:54:12 2021
 """
 
 
-def createSession(session, filepath, existing_data=False):
-    # create file paths for the session, raw and synced videos
-
-    pathList = []
-
-    recordPath = filepath / "Data"
-    recordPath.mkdir(exist_ok=True)
-
-    session.sessionPath = recordPath / session.sessionID
-    # if existing_data:
-    # session.sessionPath.mkdir(exist_ok = True)
-    # else:
-    session.sessionPath.mkdir(exist_ok=True)
-    pathList.append(session.sessionPath)
-
-    session.rawVidPath = session.sessionPath / "RawVideos"
-    session.rawVidPath.mkdir(exist_ok=True)
-    pathList.append(session.rawVidPath)
-
-    session.syncedVidPath = session.sessionPath / "SyncedVideos"
-    session.syncedVidPath.mkdir(exist_ok=True)
-    pathList.append(session.syncedVidPath)
-
-    session.calVidPath = session.sessionPath / "CalVideos"
-    session.calVidPath.mkdir(exist_ok=True)
-    pathList.append(session.calVidPath)
-
-    session.openPoseDataPath = session.sessionPath / "OpenPoseData"
-    session.openPoseDataPath.mkdir(exist_ok=True)
-    pathList.append(session.openPoseDataPath)
-
-    session.mediaPipeDataPath = session.sessionPath / "MediaPipeData"
-    session.mediaPipeDataPath.mkdir(exist_ok=True)
-    pathList.append(session.mediaPipeDataPath)
-
-    session.dlcDataPath = session.sessionPath / "DLCdata"
-    session.dlcDataPath.mkdir(exist_ok=True, parents=True)
-    pathList.append(session.dlcDataPath)
-
-    session.imOutPath = session.sessionPath / "imOut"
-    session.imOutPath.mkdir(exist_ok=True)
-    pathList.append(session.imOutPath)
-
-    session.dataArrayPath = session.sessionPath / "Data Arrays"
-    session.dataArrayPath.mkdir(exist_ok=True)
-    pathList.append(session.dataArrayPath)
-
-    # create a list of paths that is passed to the recording script
-    # pathList = [session.sessionPath,session.rawVidPath,session.syncedVidPath,session.openPoseDataPath,session.dlcDataPath,session.imOutPath, session.dataArrayPath]
-
-    config_settings, config_yaml = create_config_yaml()
-    session.config_settings = config_settings
-
-    yaml_name = session.sessionPath / "{}_config.yaml".format(session.sessionID)
-    session.yamlPath = yaml_name
-    # print(yaml_name)
-    write_config_yaml(session, config_settings, config_yaml, pathList, yaml_name)
-
-    return yaml_name
-
-
-def createSessionTxt(session, paramDict, rotDict):
-    # create a text file listing recording parameters
-    parameter_text = session.sessionPath / "sessionSettings.txt"
-    text = open(parameter_text, "w")
-    text.write("Session ID = %s\n" % (session.sessionID))
-    text.write("%s = %s\n" % ("Parameters", paramDict))
-    text.write("%s = %s\n" % ("Rotations", rotDict))
-    text.close()
 
 folder_setup = ['RawVideos','SyncedVideos','CalVideos','DataArrays','DLCData','OpenPoseData','MediaPipeData','imOut']
+
+default_parameters = {'rotations':{},'parameters':{'exposure':-5,'resWidth':640,'resHeight':480,'framerate':25,'codec':'DIVX'}}
+saved_parameters = default_parameters
+parameters_for_yaml = {'default':default_parameters,'saved':{}}
 
 
 
