### 2021-07-06 d43ee6b
Fixed unequal number of frames issue
When a camera was rotated, buffer frames were not being inserted properly into the video, which caused a difference in frames in the synced videos
