### 2021-07-06 08f3b25
Rotated videos not running from stage 2 of the pipeline bug fix
