### 2021-07-06 f5c15d7
Rotated videos not running from stage 2 of the pipeline bug fix
