+++ README.md
@@ -1,6 +1,6 @@
 # freemocap
 
-This is very much a work in progress, but a usable public release is immanent!
+This is very much a work in progress, but a usable public release is immnent!
 
 Follow the GitHub Repository and/or Join the Discord (https://discord.gg/HX7MTprYsK) for updates!
 
