### 2021-07-06 51905ff
solved the mystery of the 'cam[x]' file
the file was the pickled timestamps needed for each camera - they're now saved in the session folder
