### 2021-07-06 4435cd6
Assertion statement added to fmc_anipose.py
Assertion statement added to catch situations when no charuco board points are detected
