### 2021-07-06 e66bc13
Add pickle files and timestamp data to RawVideos folder
