+++ README.md
@@ -1,7 +1,24 @@
-# freemocap
+# Installation
 
-This is very much a work in progress, but a usable public release is imminent!
+## Prerequisates - 
+* Install Anaconda
+ 	- https://www.anaconda.com/products/individual#Downloads
+* Install CUDA
+ 	- https://developer.nvidia.com/cuda-downloads
+* Install OpenPose (Windows Portable Demo)
+  - https://github.com/CMU-Perceptual-Computing-Lab/openpose/releases/tag/v1.6.0  
+	- This can be complicated, so I'll need to add more instructions here!
+
+## Installation instructions
+- Open an Anaconda Prompt (in Windows, or any terminal on Mac/Linux) and enter the following comands
+
+`conda create -n freemocap-env python=3.7`
+`conda activate freemocap-env`
+`pip install freemocap -v`
+`ipython`
+`>>> import freemocap as fmc`
+`>>> fmc.RunMe() #this is where the magic happens `
+
+# More to come!
 
 Follow the GitHub Repository and/or Join the Discord (https://discord.gg/HX7MTprYsK) for updates!
-
-#REAL README COMING SOON!
