### 2021-07-07 #41 @NickleDave
Hi @jonmatthis we're happy you found us, thank you for your interest in pyOpenSci.

Can you say more about what the uses for `freemocap` would be?
How are you and your students applying it?
That would help clarify whether it's within scope for us.

Based on what I see on the repo, I would guess it might be more appropriate for JOSS.
We are mainly focused on tooling that supports open science. For example, if you built a package that made it easier for researchers to download motion capture datasets from a repository, or to work with different formats for annotating motion capture videos, e.g. something like a converter between COCO and Pascal VOC keypoints.
