+++ runme_fmc_staged.py
@@ -8,6 +8,9 @@ from aniposelib.boards import CharucoBoard
 
 import numpy as np
 
+
+#this is a test commit
+
 sesh = session.Session()
 
 
