+++ freemocap/__init__.py
@@ -5,5 +5,5 @@ del get_versions
 
 from freemocap import fmc_runme
 
-def RunMe(sessionID,stage,useOpenPose,useMediaPipe,useDLC,debug):
-    fmc_runme.Run(sessionID,stage,useOpenPose,useMediaPipe,useDLC,debug)
+def RunMe(sessionID,stage,useOpenPose,useMediaPipe,useDLC,dlcConfigPath,debug):
+    fmc_runme.Run(sessionID,stage,useOpenPose,useMediaPipe,useDLC,dlcConfigPath,debug)
