### 2021-07-07 #41 @lwasser
agreed. a bit more information would be very helpful @jonmatthis with that said, if you need some help in developing your package we may be able to direct you to resources. We also have a [cookiecutter](https://github.com/pyOpenSci/cookiecutter-pyopensci) however I believe that we need to make some updates to it particularly in the space of CI. So please

1. let us know more about what your package is / does so we can decide if it's in scope and then
2. if we can direct you to resources to help you build out your package we will try to do so!! We hope to add some more documentation about setting up things like CI and such in the next year.
