### 2021-07-08 5bdb90b
Aaronc fmc run_me development (#17)
* added fmc_runme

* added dlcConfigPath

* Added default parameters to runme functions
