+++ freemocap/__init__.py
@@ -3,249 +3,7 @@ from ._version import get_versions
 __version__ = get_versions()["version"]
 del get_versions
 
+from freemocap import fmc_runme
 
-
-
-from freemocap import (
-    createvideo,
-    recordingconfig,
-    runcams,
-    calibrate,
-    fmc_mediapipe,
-    fmc_openpose,
-    fmc_deeplabcut,
-    reconstruct3D,
-    playskeleton,
-    session,
-    webcamGUI
-)
-from freemocap.fmc_pyqtgraph import PlayerDockedWindow
-
-from pathlib import Path
-import os
-
-from aniposelib.boards import CharucoBoard
-
-import numpy as np
-
-def RunMe():
-
-    
-    sesh = session.Session()
-
-
-    sesh.useOpenPose = True
-    sesh.useMediaPipe = False
-    sesh.useDLC = False
-
-    if sesh.useDLC:
-        import deeplabcut as dlc
-
-        # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
-        sesh.dlcConfigPath = Path(
-            "C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml",
-        )
-
-
-    # %% Inputs to edit
-    stage = 1  # set your starting stage here (stage = 1 will run the pipeline from webcams)
-    sesh.debug = False
-
-    sesh.sessionID = (
-        ""  # fill in if you are running from Stage 2 onwards
-    )
-
-    if not sesh.sessionID:
-        dataFolder = Path.cwd() / "FreeMoCap_Data"
-
-        if dataFolder.exists():
-            subfolders = [
-                f.path for f in os.scandir(dataFolder) if f.is_dir()
-            ]  # copy-pasta from who knows where
-            sesh.sessionID = Path(
-                subfolders[-1]
-            ).stem  # grab the name of the last folder in the list of subfolders
-        else: #First time run! Make 'FreeMoCap_Data' Folder (and eventually,  prompt to download sample data)
-            dataFolder.mkdir()
-            sampleDataFolder = dataFolder / '_sample_data_folder'
-            sampleDataFolder.mkdir()
-            # prompt user if they want to download sample data
-            # if yes, download sample session folder from public URL (i.e. figshare )
-            # (also same for relevant DLC folder, which we should remove from the .git)
-            
-    board = CharucoBoard(
-        7,
-        5,
-        # square_length=1, # here, in mm but any unit works (JSM NOTE - just using '1' so resulting units will be in 'charuco squarelenghts`)
-        # marker_length=.8,
-        #  square_length = 121, #big boi charuco
-        #  marker_length = 98,
-        square_length=82,  # regular boi charuco
-        marker_length=65,
-        marker_bits=4,
-        dict_size=250,
-    ) #someday - need alter this code so that the 'squaresx' and 'squaresy' inputs are keyword arguments
-
-
-
-    # %% Initialization
-    #sesh.initialize(stage)
-    if stage ==2 or stage == 1:
-        webcamGUI.initialize(sesh,stage,board)
-    else:
-        sesh.initialize(stage)
-    # %% Stage One
-    if stage <= 1:
-        print()
-        print("Starting Video Recordings")
-        runcams.RecordCams(
-            sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs
[…]
