### 2021-07-08 c4e6b60
Aaronc fmc run_me development (#17)
* added fmc_runme

* added dlcConfigPath

* Added default parameters to runme functions
