+++ freemocap/fmc_runme.py
@@ -1,0 +1,204 @@
+from freemocap import (
+    createvideo,
+    recordingconfig,
+    runcams,
+    calibrate,
+    fmc_mediapipe,
+    fmc_openpose,
+    fmc_deeplabcut,
+    reconstruct3D,
+    playskeleton,
+    session,
+    webcamGUI
+)
+from freemocap.fmc_pyqtgraph import PlayerDockedWindow
+
+from pathlib import Path
+import os
+
+from aniposelib.boards import CharucoBoard
+
+import numpy as np
+
+def Run(sessionID=None,stage=1,useOpenPose=True,useMediaPipe=False,useDLC=True,dlcConfigPath=None,debug=False):
+    sesh = session.Session()
+
+    sesh.sessionID = sessionID
+    sesh.useOpenPose= useOpenPose
+    sesh.useMediaPipe = useMediaPipe
+    sesh.useDLC= useDLC
+    sesh.debug = debug
+  
+
+
+
+    if sesh.useDLC: 
+        import deeplabcut as dlc
+        sesh.dlcConfigPath =dlcConfigPath
+        # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
+        #sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml") 
+
+
+    if not sesh.sessionID:
+        dataFolder = Path.cwd()/'Data'
+        subfolders = [ f.path for f in os.scandir(dataFolder) if f.is_dir() ] #copy-pasta from who knows where
+        sesh.sessionID = Path(subfolders[-1]).stem #grab the name of the last folder in the list of subfolders
+
+
+
+    board = CharucoBoard(7, 5,
+                        #square_length=1, # here, in mm but any unit works (JSM NOTE - just using '1' so resulting units will be in 'charuco squarelenghts`)
+                        #marker_length=.8,
+                        #  square_length = 121, #big boi charuco
+                        #  marker_length = 98,
+                        square_length = 82,#regular boi charuco
+                        marker_length = 65,
+                        marker_bits=4, dict_size=250)
+
+
+    sesh.board = board
+    #sesh.input_stage = stage
+
+
+
+    # %% Initialization
+    #sesh.initialize(stage)
+    if stage ==2 or stage == 1:
+        webcamGUI.initialize(sesh,stage,board)
+    else:
+        sesh.initialize(stage)
+
+    # %% Stage One
+    if stage <= 1:
+        print()
+        print('Starting Video Recordings')
+        runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
+        sesh.save_session()
+    else:
+        print('Skipping Video Recording')
+
+
+    # %% Stage Two
+    if stage <= 2:
+        print()
+        print('Starting Video Syncing')
+        runcams.SyncCams(sesh, sesh.timeStampData,sesh.numCamRange,sesh.vidNames,sesh.camIDs)
+        sesh.save_session()
+    else:
+        print('Skipping Video Syncing')
+
+    # %% Stage Three
+    if stage <= 3:
+        print()
+        print('Starting Calibration')
+        sesh.cgroup, sesh.mean_charuco_fr_mar_dim = calibrate.CalibrateCaptureVolume(sesh,board)
+    else:
+        print('Skipping Calibration')
+
+    # %% Stage Four
+    if stage <= 4:
+        print('Starting Track Image Points')
[…]
