### 2021-07-08 849c67c
Create `Data` folder (with `_sample_data_folder`) on first run
