### 2021-07-08 aa77ae1
Simplified (and changed the name of) the runme script
