+++ runme_fmc_staged.py
@@ -1,216 +1,0 @@
-from freemocap import createvideo, webcamGUI, runcams, calibrate, fmc_mediapipe, fmc_openpose, fmc_deeplabcut, reconstruct3D, playskeleton, session
-from freemocap.fmc_pyqtgraph import PlayerDockedWindow
-
-from pathlib import Path
-import os
-
-from aniposelib.boards import CharucoBoard
-
-import numpy as np
-
-
-#this is a test commit
-
-sesh = session.Session()
-
-
-sesh.useOpenPose= True
-sesh.useMediaPipe = False
-sesh.useDLC=False
-
-
-
-if sesh.useDLC: 
-    import deeplabcut as dlc
-    # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
-    sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml") 
-
-
-
-# %% Inputs to edit
-stage = 1 #set your starting stage here (stage = 1 will run the pipeline from webcams)
-#6/30/21 log - AC (will be cleared when merging this with main)
- #running from State 1 works until Stage 4 openPose parsing - parseOpenPose needs to be updated, and the graphing stuff with file paths
- #running from Stage 2 does not work right now
- #running from Stage 3 for a webcam recording works until openPose parsing
- #running from Stage 3 for an external recording works until openPose parsing (will break bc we need to set numCams/numFrames manually beforehand)
-#6/31/21 log -AC
-#running from Stages 3,4,6,7 and 8 work until the end now 
-#stage 1 till the end presumably works - still need to do a full run thruogh
-#**only tested with openPose - mediaPipe probably wont work
-# -update, got stage 2 onwards to work (I think)
-# -update 2 - stage 1 running till the end works
-
-
-
-sesh.debug = False
-
-sesh.sessionID = '' #fill in if you are running from Stage 2 onwards
-
-
-
-if not sesh.sessionID:
-    dataFolder = Path.cwd()/'Data'
-    subfolders = [ f.path for f in os.scandir(dataFolder) if f.is_dir() ] #copy-pasta from who knows where
-    sesh.sessionID = Path(subfolders[-1]).stem #grab the name of the last folder in the list of subfolders
-
-
-
-board = CharucoBoard(7, 5,
-                     #square_length=1, # here, in mm but any unit works (JSM NOTE - just using '1' so resulting units will be in 'charuco squarelenghts`)
-                     #marker_length=.8,
-                    #  square_length = 121, #big boi charuco
-                    #  marker_length = 98,
-                     square_length = 82,#regular boi charuco
-                     marker_length = 65,
-                     marker_bits=4, dict_size=250)
-
-
-sesh.board = board
-#sesh.input_stage = stage
-
-
-
-# %% Initialization
-#sesh.initialize(stage)
-if stage ==2 or stage == 1:
-    webcamGUI.initialize(sesh,stage,board)
-else:
-    sesh.initialize(stage)
-
-# %% Stage One
-if stage <= 1:
-    print()
-    print('Starting Video Recordings')
-    runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
-    sesh.save_session()
[…]
