### 2021-07-08 dcaff25
Simplified (and renamed) freemocap_runme_script
