+++ freemocap/fmc_runme.py
@@ -38,11 +38,6 @@ def Run(sessionID=None,stage=1,useOpenPose=True,useMediaPipe=False,useDLC=True,d
         # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
         #sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml") 
 
-
-        dataFolder = Path.cwd()/'Data'
-        subfolders = [ f.path for f in os.scandir(dataFolder) if f.is_dir() ] #copy-pasta from who knows where
-        sesh.sessionID = Path(subfolders[-1]).stem #grab the name of the last folder in the list of subfolders
-
     if dataFolder.exists():
         subfolders = [f.path for f in os.scandir(dataFolder) if f.is_dir()]  # copy-pasta from who knows where
         sesh.sessionID = Path(subfolders[-1]).stem  # grab the name of the last folder in the list of subfolders
