+++ README.md
@@ -1,5 +1,20 @@
 # This is all very much a work in progress! More to come!
 
+## Installation instructions (but FYI - This still isn't *really* in a state that's usable for outside users yet 😅 We're working on it though, stay tuned!)
+- Open an Anaconda Prompt (in Windows, or any terminal on Mac/Linux) and enter the following comands
+
+`conda create -n freemocap-env python=3.7`
+
+`conda activate freemocap-env`
+
+`pip install freemocap -v`
+
+`ipython`
+
+```Python Console
+import freemocap as fmc
+fmc.RunMe() #this is where the magic happens. Also, as of 2021-07-08, is likely going to be pretty crashy/buggy
+```
 
 https://user-images.githubusercontent.com/15314521/124694557-8069ea00-deaf-11eb-9328-3be27a4b1ea4.mp4
 
@@ -28,25 +43,6 @@ https://user-images.githubusercontent.com/15314521/124694557-8069ea00-deaf-11eb-
 	cv2.imwrite('charuco_board_image.png',charuco_board_image)
 	
 	```
-
-## Installation instructions (but FYI - This still isn't *really* in a state that's usable for outside users yet 😅 We're working on it though, stay tuned!)
-- Open an Anaconda Prompt (in Windows, or any terminal on Mac/Linux) and enter the following comands
-
-`conda create -n freemocap-env python=3.7`
-
-`conda activate freemocap-env`
-
-`pip install freemocap -v`
-
-`ipython`
-
-```Python Console
-import freemocap as fmc
-```
-
-```Python Console
-fmc.RunMe() #this is where the magic happens. Also, as of 2021-07-08, is likely going to be pretty crashy/buggy
-```
 
 
 
