+++ README.md
@@ -34,7 +34,7 @@ https://user-images.githubusercontent.com/15314521/124694557-8069ea00-deaf-11eb-
 	```
 	import cv2
 	
-	aruco_dict = aruco.Dictionary_get(cv2.aruco.DICT_4X4_250) #note `cv2.aruco` can be installed via `pip install opencv-contrib-python`
+	aruco_dict = cv2.aruco.Dictionary_get(cv2.aruco.DICT_4X4_250) #note `cv2.aruco` can be installed via `pip install opencv-contrib-python`
 	
 	board = cv2.aruco.CharucoBoard_create(7, 5, 1, .8, aruco_dict)
 	
