### 2021-07-09 775e792
GUI updates with folder path
Added ability to choose basepath for data saving to the GUIs. changed the 'parameters.yaml' to 'user_preferences.yaml'
