+++ freemocap/recordingconfig.py
@@ -5,11 +5,11 @@ Created on Thu Mar 11 11:54:12 2021
 @author: Rontc
 """
 
-
+from pathlib import Path
 
 folder_setup = ['RawVideos','SyncedVideos','CalVideos','DataArrays','DLCData','OpenPoseData','MediaPipeData','imOut']
 
-default_parameters = {'rotations':{},'parameters':{'exposure':-5,'resWidth':640,'resHeight':480,'framerate':25,'codec':'DIVX'}}
+default_parameters = {'path_to_save': str(Path.cwd()),'rotations':{},'parameters':{'exposure':-5,'resWidth':640,'resHeight':480,'framerate':25,'codec':'DIVX'}}
 saved_parameters = default_parameters
 parameters_for_yaml = {'default':default_parameters,'saved':{}}
 
