+++ freemocap/webcam/recordGUI.py
@@ -5,10 +5,11 @@ Created on Wed Mar  3 12:18:46 2021
 @author: Rontc
 """
 import tkinter as tk
-from tkinter import Tk, Label, Button, Frame, Listbox, Entry
+from tkinter import Tk, Label, Button, Frame, Listbox, Entry, filedialog
 from freemocap.webcam import checkcams
 import pickle
 import datetime
+from pathlib import Path
 
 
 class recordGUI:
@@ -77,6 +78,7 @@ class SettingsGUI:
         rotation_dictionary,
         sessionID_in,
         task,
+        current_path_to_save,
     ):
         self.master = master
         self.cam_inputs = cam_inputs
@@ -84,6 +86,7 @@ class SettingsGUI:
         self.rotation_dictionary = rotation_dictionary
         self.sessionID_in = sessionID_in
         self.task = task
+        self.currentPath = current_path_to_save
         # check if any previous parameters exist
         if self.parameter_dictionary is not None:
             existing_parameters = True
@@ -96,6 +99,9 @@ class SettingsGUI:
         # create frames to organize rotation inputs, resolution inputs, and parameter inputs
         topFrame = Frame(self.master)
         topFrame.pack(side=tk.TOP)
+        if self.task == 'record':
+            recordingSettingsFrame = Frame(self.master)
+            recordingSettingsFrame.pack()
         rotLabelFrame = Frame(self.master)
         rotLabelFrame.pack()
         resFrame = Frame(self.master)
@@ -103,6 +109,8 @@ class SettingsGUI:
         parametersFrame = Frame(self.master)
         parametersFrame.pack()
         bottomFrame = Frame(self.master).pack(side=tk.BOTTOM)
+
+      
 
         taskText = "Current Task: " + self.task
         taskLabel = Label(topFrame, text=taskText, font="bold")
@@ -151,11 +159,31 @@ class SettingsGUI:
         # ---SessionID entry- default is the date/time
         if self.task == "record":
             sessionText = "Change SessionID if desired"
-            sessionLabel = Label(parametersFrame, text=sessionText).pack()
-            self.sessionIDEntry = Entry(parametersFrame)
+            sessionLabel = Label(recordingSettingsFrame, text=sessionText)
+            sessionLabel.pack()
+
+            self.sessionIDEntry = Entry(recordingSettingsFrame)
             self.sessionIDEntry.pack()
             self.sessionIDEntry.insert(0, self.sessionID_in)
 
+            pathText = 'Choose where to create your Freemocap_Data folder:'
+            pathLabel = Label(recordingSettingsFrame, text=pathText)
+            pathLabel.pack()
+
+            text = 'Current file path: ' + self.currentPath
+            self.currentPathText = tk.StringVar()
+            self.currentPathText.set(text)
+            currentPathLabel = Label(recordingSettingsFrame,textvariable = self.currentPathText)
+            currentPathLabel.pack()
+
+            self.path_changed = False #will be set to True if the user changes the path with the file dialog
+
+            changepath_button = Button(recordingSettingsFrame,text = 'Change File Path',command = self.openFileDialog)
+            changepath_button.pack(side = 'left')
+            
+            #savedefault_button = Button(recordingSettingsFrame,text = 'Save as Default Path')
+            #savedefault_button.pack(side = 'left')
+            
         # ---If previous parameters were recieved, insert them into the entry boxes
         if existing_parameters == True:
 
@@ -168,6 +196,13 @@ class SettingsGUI:
         # ---Submit button to hit when things are finalized
         self.sub_button = Button(bottomFrame, text="Submit", command=self.Submit)
[…]
