+++ freemocap/playskeleton.py
@@ -37,6 +37,7 @@ def ReplaySkeleton_matplotlib(
     useDLC=False,
 ):
 
+    startFrame = 150
     
     mean_charuco_fr_mar_dim = np.load(session.dataArrayPath/'charuco_3d_points.npy')
 
@@ -50,9 +51,8 @@ def ReplaySkeleton_matplotlib(
 
     if useDLC:
         dlc_fr_mar_dim = np.load(session.dataArrayPath / "deepLabCut_3d.npy")
-        dlc0 = np.squeeze(dlc_fr_mar_dim[:, 0, :])
-        dlc1 = np.squeeze(dlc_fr_mar_dim[:, 1, :])
-        dlc2 = np.squeeze(dlc_fr_mar_dim[:, 2, :])
+        dlcData_nCams_nFrames_nImgPts_XYC = np.load(session.dataArrayPath / "deepLabCutData_2d.npy")
+
         ballTrailLen = 4
 
     if useOpenPose:
@@ -64,6 +64,8 @@ def ReplaySkeleton_matplotlib(
         camImgPathList[cam] = list(sorted(Path(imgPathList[cam]).glob('*.png')))
         #session.numFrames = len(camImgPathList[cam]) #will need to perhaps put a check in on whether numFrames between cameras are the same
     
+    #get list of synched video paths
+    syncedVidPathList = list(sorted(session.syncedVidPath.glob('*.mp4')))
 
 # define Skeleton connections 
 # head = [17, 15, 0, 16, 18]
@@ -89,11 +91,25 @@ def ReplaySkeleton_matplotlib(
 
     fig = plt.figure(figsize=(11, 8))
 
-    if vidType == 0:
+    if vidType == 0: #no videos
         axMain = fig.add_subplot(111, projection="3d")
-    elif vidType == 1:
+    elif vidType == 1: #one video
+        numCamSubPlots = 1
         axMain = fig.add_subplot(121, projection="3d")
         axCam1 = fig.add_subplot(122)
+
+    #make a list of video capture objects to load images from as needed for later plot
+    for camNum in range(numCamSubPlots):
+        vidCapObj_list = []
+        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[camNum+1])))
+
+    #read over the frames preceeding startFrame, so that the video will be at the right place in time for the plotting loop
+    #(I am doing this b/c I can't figure out how to set the capture object to start reading AT startframe....)
+    # SLOPPY TECH DEBT
+    for fr in range(startFrame):
+        for thisVidObj in enumerate(vidCapObj_list):
+            success, image = thisVidObj[1].read()
+            assert(success, '{} - Failed to Read Frame'.format(syncedVidPathList[thisVidObj[0]].stem))
 
     # fig.tight_layout(h_pad = 5)
     axMain.view_init(azim=azimuth, elev=elevation)
@@ -116,9 +132,8 @@ def ReplaySkeleton_matplotlib(
         sk_z = skel_fr_mar_dim[fr, :, 2]  # skeleton z data
 
         axMain.scatter3D(sk_x, sk_y, sk_z, marker=".", color="k", s=4.0)
-        axMain.plot(
-            sk_x[head], sk_y[head], sk_z[head], linestyle="-", color="g", linewidth=1.0
-        )
+        axMain.plot(sk_x[head], sk_y[head], sk_z[head], linestyle="-", color="g", linewidth=1.0)
+        
         axMain.plot(
             sk_x[spine],
             sk_y[spine],
@@ -240,318 +255,18 @@ def ReplaySkeleton_matplotlib(
             color="b",
[…]
