+++ freemocap_runme_script.py
@@ -1,3 +1,5 @@
 import freemocap as fmc
+from pathlib import Path
 
-fmc.RunMe()
+dlcConfigPath = Path(r"C:\Users\jonma\Dropbox\DLC_Models\PinkGreenRedJugglingBalls-JSM-2021-05-31\config.yaml")
+fmc.RunMe(stage=4, openPoseDummyRun=True, dlcConfigPath=dlcConfigPath) 
