### 2021-07-11 8993e68
working with matplot lib 'Animation' methods
