### 2021-07-13 ca71613
Updating data path choices
Lets user choose data path for rerunning recordings, bug fixes, overall workflow changes
