### 2021-07-13 c24c3d2
Updating data path choices
Lets user choose data path for rerunning recordings, bug fixes, overall workflow changes
