+++ freemocap/session.py
@@ -33,10 +33,10 @@ class Session: #self like "recording self"
 
     def start_session(self,paramDict,rotDict):
         #start a session, create all the file paths necessary, and create a session dictionary to save settings
-        recordPath = self.basePath/'Data'
-        recordPath.mkdir(exist_ok= True)
+        dataFolderPath = self.basePath/self.dataFolderName
+        dataFolderPath.mkdir(exist_ok= True)
 
-        self.sessionPath = recordPath/self.sessionID
+        self.sessionPath = dataFolderPath/self.sessionID
         self.sessionPath.mkdir(exist_ok=True)
 
         #create all the session filepaths and settings - create and load them into a dictionary
@@ -129,10 +129,10 @@ class Session: #self like "recording self"
     def initialize(self,stage):
         #load all session settings back into the session class for this run-through of the code
         
-        recordPath = self.basePath/'Data' #create a Data folder in the filepath if none exists yet
-        recordPath.mkdir(exist_ok= True)
+        #recordPath = self.basePath/'Data' #create a Data folder in the filepath if none exists yet
+        #recordPath.mkdir(exist_ok= True)
 
-        self.sessionPath = recordPath/self.sessionID
+        self.sessionPath = self.dataFolderPath/self.sessionID
         self.sessionPath.mkdir(exist_ok=True)
 
         self.session_yaml_path = self.sessionPath/'{}_config.yaml'.format(self.sessionID)
