### 2021-07-13 5011f6b
Updating data path choices
Lets user choose data path for rerunning recordings, bug fixes, overall workflow changes
