### 2021-07-13 5970a37
some comments and changes for efficiency
