### 2021-07-13 3c5a0b0
Updating skeleton animation code
gonna use 'FuncAnimation' like a grown up!
