+++ freemocap_runme_script.py
@@ -2,4 +2,4 @@ import freemocap as fmc
 from pathlib import Path
 
 dlcConfigPath = Path(r"C:\Users\jonma\Dropbox\DLC_Models\PinkGreenRedJugglingBalls-JSM-2021-05-31\config.yaml")
-fmc.RunMe(stage=4, openPoseDummyRun=True, dlcConfigPath=dlcConfigPath) 
+fmc.RunMe(sessionID = 'sesh_21-07-08_131030', stage=7, openPoseDummyRun=True, useDLC = False, dlcConfigPath=dlcConfigPath) 
