### 2021-07-13 7209f36
Updating skeleton animation code
gonna use 'FuncAnimation' like a grown up!
