### 2021-07-13 4652279
stable skeleton playback (not yet * - a e s t h e t i c - *)
