### 2021-07-13 489be92
reorganizing where user preferences are decided
user preferences decided at beginning of runme

session class has a 'save_user_preferences' utility function
