+++ freemocap/session.py
@@ -183,7 +183,10 @@ class Session: #self like "recording self"
         self.imOutPath = session_paths['imOut']
         self.dataArrayPath = session_paths['DataArrays']
 
-        
+    def save_user_preferences(self,preferences):
+        preferences_yaml = YAML()
+        preferences_yaml.dump(preferences,self.preferences_path)
+
 
         f = 2
 
