### 2021-07-13 8e525ed
reorganizing where user preferences are decided
user preferences decided at beginning of runme

session class has a 'save_user_preferences' utility function
