+++ freemocap_test.py
@@ -1,0 +1,14 @@
+import freemocap
+
+# sessionID = ''
+# stage = 1
+# useOpenPose = True
+# useMediaPipe = False
+# useDLC = False
+# debug = False
+
+
+# freemocap.RunMe(sessionID,stage,useOpenPose,useMediaPipe,useDLC,debug)
+
+freemocap.RunMe( stage = 1, useDLC=True,setDataPath=False)
+
