### 2021-07-15 a6b9cf7
Hideous but extant video subplots
