### 2021-07-19 64f1270
Clean up animation and add 'recordVid' bool
