+++ freemocap/play_skeleton_animation.py
@@ -1,0 +1,504 @@
+
+"""
+============
+██████  ██       █████  ██    ██     ███████ ██   ██ ███████ ██      ███████ ████████  ██████  ███    ██      █████  ███    ██ ██ ███    ███  █████  ████████ ██  ██████  ███    ██ 
+██   ██ ██      ██   ██  ██  ██      ██      ██  ██  ██      ██      ██         ██    ██    ██ ████   ██     ██   ██ ████   ██ ██ ████  ████ ██   ██    ██    ██ ██    ██ ████   ██ 
+██████  ██      ███████   ████       ███████ █████   █████   ██      █████      ██    ██    ██ ██ ██  ██     ███████ ██ ██  ██ ██ ██ ████ ██ ███████    ██    ██ ██    ██ ██ ██  ██ 
+██      ██      ██   ██    ██             ██ ██  ██  ██      ██      ██         ██    ██    ██ ██  ██ ██     ██   ██ ██  ██ ██ ██ ██  ██  ██ ██   ██    ██    ██ ██    ██ ██  ██ ██ 
+██      ███████ ██   ██    ██        ███████ ██   ██ ███████ ███████ ███████    ██     ██████  ██   ████     ██   ██ ██   ████ ██ ██      ██ ██   ██    ██    ██  ██████  ██   ████                                                                                                                                                                                                                                                                                                                                                                         
+============
+Font - ANSI Regular - https://patorjk.com/software/taag/#p=display&f=ANSI%20Regular&t=Play%20Skeleton%20Animation
+============
+Originally based on tutorial from  - https://matplotlib.org/2.1.2/gallery/animation/simple_3danim.html
+
+Create a matplotlib animation from a FreeMoCap recording session.
+"""
+
+
+################################################################################################################
+################################################################################################################
+###
+### ██ ███    ███ ██████   ██████  ██████  ████████ 
+### ██ ████  ████ ██   ██ ██    ██ ██   ██    ██    
+### ██ ██ ████ ██ ██████  ██    ██ ██████     ██    
+### ██ ██  ██  ██ ██      ██    ██ ██   ██    ██    
+### ██ ██      ██ ██       ██████  ██   ██    ██    
+###
+################################################################################################################
+################################################################################################################                                                        
+                                                        
+
+import numpy as np
+
+import matplotlib.pyplot as plt
+import mpl_toolkits.mplot3d.axes3d as p3
+import matplotlib.animation as animation
+from matplotlib.widgets import Slider
+import cv2 
+
+import copy
+from pathlib import Path
+
+#RICH CONSOLE STUFF
+from rich import pretty
+pretty.install() #makes all print statement output pretty
+from rich import inspect
+from rich.console import Console
+console = Console()  
+from rich.traceback import install as rich_traceback_install
+from rich.markdown import Markdown
+
+#colors from Taylor Davis branding - 
+humon_dark = np.array([37, 67, 66])/255
+humon_green = np.array([53, 93, 95])/255
+humon_blue = np.array([164, 211, 217])/255
+humon_red = np.array([217, 81, 87])/255
+
+
+def PlaySkeletonAnimation(
+    session=None,
+    vidType=1,
+    startFrame=0,
+    azimuth=-90,
[…]
