### 2021-07-19 a3867de
Clean up animation and add 'recordVid' bool
