+++ freemocap/dev_scratchpad/__init__.py
@@ -1,0 +1,1 @@
+from play_skeleton_animation.py import PlaySkeletonAnimation
