+++ freemocap/dev_scratchpad/play_skeleton_animation.py
@@ -1,307 +1,0 @@
-
-"""
-============
-3D animation
-============
-
-An animated plot in 3D.
-from - https://matplotlib.org/2.1.2/gallery/animation/simple_3danim.html
-"""
-import numpy as np
-
-import matplotlib.pyplot as plt
-import mpl_toolkits.mplot3d.axes3d as p3
-import matplotlib.animation as animation
-from matplotlib.widgets import Slider
-import cv2 
-
-import copy
-from pathlib import Path
-
-#RICH CONSOLE STUFF
-from rich import pretty
-pretty.install() #makes all print statement output pretty
-from rich import inspect
-from rich.console import Console
-console = Console()  
-from rich.traceback import install as rich_traceback_install
-from rich.markdown import Markdown
-
-
-
-def PlaySkeletonAnimation(
-    session=None,
-    vidType=1,
-    startFrame=40,
-    azimuth=-90,
-    elevation=-80,
-    useOpenPose=True,
-    useMediaPipe=False,
-    useDLC=False,
-    ):
-  
-    def update_figure(frameNum):
-        """ 
-        Called by matplotlib animator for each frame.
-        """
-        
-        skel_dottos = matplotlib_artist_objs['skel_dottos'] 
-        skel_trajectories = figure_data['skel_trajectories|mar|fr_dim']
-
-
-        #function to update the lines for each body segment
-        def update_skel_segments(key):       
-            """ 
-            updates the Artist of each body segment with the current frame data.
-            """       
-            dict_of_segments_idxs = dict_of_openPoseSegmentIdx_dicts[key]
-            dict_of_artists = matplotlib_artist_objs[key] 
-
-            for thisItem in dict_of_segments_idxs.items():
-                segName = thisItem[0]
-                segArtist = dict_of_artists[segName]
-                segArtist.set_data((skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 0], skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 1]))
-                segArtist.set_3d_properties(skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 2])                
-            
-        update_skel_segments('body')
-        update_skel_segments('rHand')
-        update_skel_segments('lHand')
-        update_skel_segments('face')
-
-        #Plots the dots!
-        marNum = -1
-        for thisSkelDotto, thisTraj in zip(skel_dottos,skel_trajectories):
-            marNum+=1
-            # NOTE: there is no .set_data() for 3 dim data...
-            thisSkelDotto.set_data(thisTraj[ frameNum-1, 0:2])
-            thisSkelDotto.set_3d_properties(thisTraj[ frameNum-1, 2])
-
-        for vidNum, thisVidArtist in enumerate(vidAristList):
-            success, image = vidCapObjList[vidNum].read()
-            thisVidArtist.set_array(cv2.cvtColor(image, cv2.COLOR_BGR2RGB))
-
-
-        # animSlider.set_val(val=frameNum)
-
-    ##
-    ##
-    ## Create the figure for the plot
-    ##
-    ##
-
-    fig = plt.figure(dpi=150)
-    plt.ion()
-
-    ###
-    ###
-    ### Attaching 3D axis to the figure
-    ###
-    ###
-    
-    ax3d = p3.Axes3D(fig)
[…]
