### 2021-07-19 dc26473
Reworked data handler and Maya tools. Now Python 3 native
