### 2021-07-19 d551fd9
Reworked data handler and Maya tools. Now Python 3 native
