+++ freemocap/fmc_trackdata_handler.py
@@ -1,0 +1,236 @@
+import numpy as np
+from pathlib import Path
+
+FILE_SUFFIX = ".fmcData"
+
+class FmcTracDataHandler(object):
+    """
+    # data structure:
+    {
+    "header":{
+        "version": str(),
+        "tags": list(),
+        "author": str(),
+        "export_date": datetime(),
+        "camera_count": int(),
+        "license": str(),
+        "calibration_obj": class,
+        },
+    "content":{
+        str(actor_name): {
+            "sample_count": int(),
+            "tracking_points": {
+                str(tracking_point_name): {"parents": list(), samples: list()}
+            }
+        }
+    },
+    }
+    """
+
+    def __init__(self):
+        self.data = {"header":{}, "content":{}}
+
+    def has_actor(self, actor_name):
+        return actor_name in self.data["content"]
+
+    def has_trackingpoint_name(self, actor_name, point_name):
+        return self.has_actor(actor_name) and point_name in self.data["content"][actor_name]["tracking_points"]
+
+    def set_point_name_mapping(self, point_name_mapping):
+        self.point_name_mapping = point_name_mapping
+
+    def set_parent_mapping(self, parent_mapping):
+        self.parent_mapping = parent_mapping
+
+    def init_actor(self, actor_name):
+        if not self.has_actor(actor_name):
+            self.data["content"][actor_name] = {}
+        else:
+            raise Exception("Actor already exist in dataset. Please remove it before re-adding")
+
+    def init_tracking_point_name(self, actor_name, point_name):
+        if not self.has_actor(actor_name):
+            self.init_actor(actor_name)
+        if not "tracking_points" in self.data["content"][actor_name]:
+            self.data["content"][actor_name]["tracking_points"] = {}
+        self.data["content"][actor_name]["tracking_points"][point_name] = {}
+
+    def set_actor_data(self, actor_name, data_label, data):
+        if not self.has_actor(actor_name):
+            self.init_actor(actor_name)
+        self.data["content"][actor_name][data_label] = data
+
+
+    def set_tracking_point_samples(self, actor_name, point_name, samples):
+        if not self.has_trackingpoint_name(actor_name, point_name):
+            self.init_tracking_point_name(actor_name, point_name)
+        self.data["content"][actor_name]["tracking_points"][point_name]["samples"] = samples
+
+    def set_tracking_point_parents(self, actor_name, point_name, parents):
+        if not self.has_trackingpoint_name(actor_name, point_name):
+            self.init_tracking_point_name(actor_name, point_name)
+        self.data["content"][actor_name]["tracking_points"][point_name]["parents"] = parents
+
+    def set_actor_sample_count(self, actor_name, sample_count):
+        if self.has_actor(actor_name):
+            self.data["content"][actor_name]["sample_count"] = sample_count
+        else:
+            raise ValueError("Object does not contain an actor called: %s" % actor_name)
+
+    def set_header(key, value):
+        self.data["header"][key] = value
+
+    def set_version(self, version):
+        self.set_header("version", version)
+
+    def set_tags(self, tag_list):
+        self.set_header("tags", tag_list)
+
+    def set_author(self, author):
+        self.set_header("author", author)
+
+    def set_date(self, datetime_obj):
+        self.set_header("export_date", datetime_obj)
+
+    def set_camera_count(self, camera_count):
+        self.set_header("camera_count", camera_count)
+
+    def set_license(self, license_string):
+        self.set_header("license", license_string)
+
+    def set_calibration_obj(self, calibration_obj):
+        self.set_header("calibration_obj", calibration_obj)
+
+    def import_actor_raw_data(self, actor_label, raw_data):
+        mapped_data = self.map_point_names(raw_data)
+        parent_mapping = self.get_parent_mapping()
+
+        sample_count = False
[…]
