+++ freemocap/datastruct_formatter.py
@@ -1,334 +1,0 @@
-import pickle
-import numpy as np
-from pathlib import Path
-
-
-
-
-class DataStructure(object):
-    """
-    Main class containing mmcap dataset
-
-    Attributes:
-        header: Dictionary with all header information
-        content: List of Actor objects
-    """
-    def __init__(self):
-        self.header = {
-                "version": False, #str()
-                "tags": False, #list(),
-                "author": False, #str(),
-                "export_date": False, #datetime(),
-                "sample_rate": False, #float(),
-                "camera_count": False, #int(),
-                "license": False, #str(),
-                "calibration_obj": False, #CalibrationObj(),
-            }
-        self.content = [] # list of Actor objects
-
-    def set_version(self, version):
-        self.header["version"] = version
-
-    def set_tags(self, tag_list):
-        self.header["tags"] = tag_list
-
-    def set_author(self, author):
-        self.header["author"] = author
-
-    def set_date(self, datetime_obj):
-        self.header["export_date"] = datetime_obj
-
-    def set_camera_count(self, camera_count):
-        self.header["camera_count"] = camera_count
-
-    def set_license(self, license_string):
-        self.header["license"] = license_string
-
-    def set_calibration_obj(self, calibration_obj):
-        self.header["calibration_obj"] = calibration_obj
-
-    def add_actor(self, actor_obj):
-        if isinstance(actor_obj, Actor):
-            self.content.append(actor_obj)
-        else:
-            raise TypeError("actor_obj should be of type Actor")
-
-    def get_header(self):
-        return self.header
-
-    def get_actors(self):
-        return self.content
-
-    def get_actor(self, actor_name):
-        for actor in self.content:
-            if actor.get_name() == actor_name:
-                return actor
-        raise ValueError(f"Actor %s not found in content {actor_name}")
-
-    def list_actor_names(self):
-        actor_list = []
-        for actor in self.content:
-            actor_list.append(actor.get_name())
-        return actor_list
-
-    def print_stats(self):
-        print("\n### Header")
-        header = self.get_header()
-        for attr in header:
-            print(attr, ":", header[attr])
-
-        print("\n### Content")
-        actor_name_list = self.list_actor_names()
-        print("Contains %s actor(s): %s" %(len(actor_name_list), actor_name_list))
-
-        for actor_name in actor_name_list:
-            actor = self.get_actor(actor_name)
-
-            act_type = actor.get_type()
-            sample_rate = actor.get_sample_rate()
-            sample_length = actor.get_length()
-            point_names = actor.list_points()
-
-            print(f"\nActor: '{actor_name}'\nsample rate: {sample_rate}\nsample length: {sample_length}\nnumber of points: {len(point_names)}")
-
-    def get_as_dict(self):
-        # WARNING: Overrides "content" attribute in place
-        actors = self.get_actors()
-        out_dict = vars(self)
-        out_dict["content"] = []
-
-        for actor in actors:
-            out_dict["content"].append(vars(actor))
-        return out_dict
-
-
-class CalibrationObj(object):
-    """
-    Handling the calibration object which has yet to be defined
-    """
-    def __init__(self):
-        pass
-
-
-class Actor(object):
-    """
[…]
