+++ requirements.txt
@@ -17,3 +17,4 @@ tf_slim>=1.1.0
 deeplabcut[gui]>=2.2rc3
 ruamel.yaml>=0.15.0
 typed-ast<1.5,>=1.4.0
+ffmpeg
