+++ freemocap/play_skeleton_animation.py
@@ -30,11 +30,13 @@ Create a matplotlib animation from a FreeMoCap recording session.
 
 import numpy as np
 
+import matplotlib as mpl
 import matplotlib.pyplot as plt
 import mpl_toolkits.mplot3d.axes3d as p3
 import matplotlib.animation as animation
 from matplotlib.widgets import Slider
 import cv2 
+from scipy.signal import savgol_filter
 
 import copy
 from pathlib import Path
@@ -51,8 +53,8 @@ from rich.markdown import Markdown
 #colors from Taylor Davis branding - 
 humon_dark = np.array([37, 67, 66])/255
 humon_green = np.array([53, 93, 95])/255
-humon_blue = np.array([164, 211, 217])/255
-humon_red = np.array([217, 81, 87])/255
+humon_blue = np.array([14, 90, 253])/255
+humon_red = np.array([217, 61, 67])/255
 
 
 def PlaySkeletonAnimation(
@@ -60,7 +62,7 @@ def PlaySkeletonAnimation(
     vidType=1,
     startFrame=0,
     azimuth=-90,
-    elevation=-71,
+    elevation=-61,
     numCams=4,
     useCams = None,
     useOpenPose=True,
@@ -88,24 +90,43 @@ def PlaySkeletonAnimation(
             dlc_dottos = matplotlib_artist_objs['dlc_dottos'] 
             dlc_trajectories = figure_data['dlc_trajectories|mar|fr_dim']
 
-        #function to update the lines for each body segment
-        def update_skel_segments(key):       
+        #function to update the lines for each 3d body segment
+        def update_3d_skel_segments(key, data_fr_mar_dim):       
             """ 
             updates the Artist of each body segment with the current frame data.
-            """       
-            dict_of_segments_idxs = dict_of_openPoseSegmentIdx_dicts[key]
+            """    
+            try:   
+                dict_of_segments_idxs = dict_of_openPoseSegmentIdx_dicts[key]
+            except: 
+                dict_of_segments_idxs = dict_of_dlcSegmentIdx_dicts[key]
+
             dict_of_artists = matplotlib_artist_objs[key] 
 
             for thisItem in dict_of_segments_idxs.items():
                 segName = thisItem[0]
                 segArtist = dict_of_artists[segName]
-                segArtist.set_data((skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 0], skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 1]))
-                segArtist.set_3d_properties(skel_fr_mar_dim[frameNum-1, dict_of_segments_idxs[segName], 2])                
+                segArtist.set_data((data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 0], data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 1]))
+                segArtist.set_3d_properties(data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 2])                
             
-        update_skel_segments('body')
-        update_skel_segments('rHand')
-        update_skel_segments('lHand')
-        update_skel_segments('face')
+        update_3d_skel_segments('body', skel_fr_mar_dim)
+        update_3d_skel_segments('rHand', skel_fr_mar_dim)
+        update_3d_skel_segments('lHand', skel_fr_mar_dim)
+        update_3d_skel_segments('face', skel_fr_mar_dim)
+
[…]
