### 2021-07-22 e641322
19: Video naming convention is 1-indexed but camera ID's are zero-indexed
