+++ .travis.yml
@@ -13,6 +13,9 @@ env:
     # Doctr deploy key for jonmatthis/freemocap
     - secure: "XP8R3/9jzF71gUJVzg49XcPDKc5lvME0X3iOFf3dKii8IVyvBFM0bc5JBVIPwV7nVJQTHCX0AGEdZSAjT85ie+EoTSm1smD1tHx0gfPJRXfHm6AXXQch7l4kDs9LpEY8S30DVgXiQdgQKTlcl8bwNl4gCpWLG8mclJoQ9hF4OyTkPBsutUjcIiB2Jp8v02CBqXNK0/dQ7dzZzLuu8UK4IwLp80yLK7WVEQGBGF/h+X4eWelRgxpXDHbtLf/OF/YunJYm6bNNXBZnaBHSL9xbmztvn7UWB9gCoJ/9afj2enGesQqlO1pHB7Zmsmz3KsNQjPADQWdd1CXN9nGGceTn7CP0Mfph6hIFnD38nBD7OWwnV3jCKpBCCYXGmc+qwBm0Drbx/wnZOiOxYS3ih+aD4Zkh9FocPwYoKq+U9/urHxtmOxjy0N+aDR3WlzjdAbTybgyX15KCjYpXHSSHG3IIP7NuFnlyRN86XKNLREsFoZvhA0kGqKrZnPojC0OZYyZ7tJl9Plho2GSt66MArJYg8npH5ELV8S/UehlG9PGMZiyvaih1Y5gJqYWddRMBuVy4y4jPQD0FIc53qJeuXBzWLUH2TyJ97zzX5Ri9gwl8Ayr+60XjWel9mNowHLsN18IkqwLfNXElCRgli3L/1iB+LEAmDlpLNp5DnoHzM0qcBKw="
 
+before_install:
+  - sudo apt-get -y install libgtk-3-dev
+
 install:
   # The versions of pip and numpy that come pre-installed are often too old and
   # can cause problems. Upgrade them.
