### 2021-07-22 ed38a66
Try to install wPython from Ubuntu package
