### 2021-07-22 1a8bbf5
Install wxPython wheels recommended way
Using 18.04, because it is the last Ubuntu with wxPython 4.0.7 wheels
https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04/
