### 2021-07-22 17accbd
FreeMoCap uses wxPython 4.0 for some reason
