### 2021-07-22 df74d3f
Fixing ImportError: Cannot load backend 'WXAgg'
```
______________________ ERROR collecting freemocap_test.py ______________________
ImportError while importing test module '/home/travis/build/jonmatthis/freemocap/freemocap_test.py'.
Hint: make sure your test modules/packages have valid Python names.
Traceback:
../../../virtualenv/python3.7.6/lib/python3.7/importlib/__init__.py:127: in import_module
    return _bootstrap._gcd_import(name[level:], package, level)
freemocap_test.py:13: in <module>
    freemocap.RunMe( stage = 1, useDLC=True,setDataPath=False)
freemocap/__init__.py:9: in RunMe
  
[…]
