+++ freemocap/fmc_runme.py
@@ -52,11 +52,6 @@ def Run(sessionID=None,
     #    sesh.dlcConfigPath = dlcConfigPath
 
 
-<<<<<<< HEAD
-    if sesh.useDLC and stage<=4:
-        import deeplabcut as dlc
-        sesh.dlcConfigPath =dlcConfigPath
-=======
 
     #%% load user preferences if they exist, create a new preferences yaml if they don't
     here = Path(__file__).parent
@@ -87,7 +82,6 @@ def Run(sessionID=None,
         sesh.preferences['saved']['dlc_config_paths'] = dlc_config_paths
         sesh.save_user_preferences(sesh.preferences)
 
->>>>>>> main
         # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
         #sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml") 
     if stage > 1:
@@ -116,15 +110,6 @@ def Run(sessionID=None,
         if not dataFolder.exists():
             raise FileNotFoundError('No data folder located at: ' + str(dataFolder))
 
-<<<<<<< HEAD
-    dataFolder = Path.cwd()/'Data'
-    if dataFolder.exists():
-        if not sesh.sessionID: #if user has not provided a sessionID, use the most recent session
-            subfolders = [f.path for f in os.scandir(dataFolder) if f.is_dir()]  # copy-pasta from who knows where
-            sesh.sessionID = Path(subfolders[-1]).stem  # grab the name of the last folder in the list of subfolders
-    else: #First time run! Make 'FreeMoCap_Data' Folder (and eventually,  prompt to download sample data)
-        dataFolder.mkdir()
-=======
         if sesh.sessionID == None:    
             subfolders = [f.path for f in os.scandir(sesh.dataFolderPath) if f.is_dir()]  # copy-pasta from who knows where
             sesh.sessionID = Path(subfolders[-1]).stem  # grab the name of the last folder in the list of subfolders
@@ -157,7 +142,6 @@ def Run(sessionID=None,
     #    dataFolder.mkdir()
     #    sampleDataFolder = dataFolder / '_sample_data_folder'
     #    sampleDataFolder.mkdir()
->>>>>>> main
 
     board = CharucoBoard(7, 5,
                         #square_length=1, # here, in mm but any unit works (JSM NOTE - just using '1' so resulting units will be in 'charuco squarelenghts`)
@@ -228,21 +212,11 @@ def Run(sessionID=None,
             for vid in sesh.syncedVidPath.glob('*.mp4'):
                 sesh.syncedVidList.append(str(vid))
             
-<<<<<<< HEAD
-            dlc.analyze_videos(str(sesh.dlcConfigPath),sesh.syncedVidList, destfolder= sesh.dlcDataPath, save_as_csv=True) 
-
-            sesh.dlcData_nCams_nFrames_nImgPts_XYC = fmc_deeplabcut.parseDeepLabCut(sesh)
-            np.save(sesh.dataArrayPath / "deepLabCutData_2d.npy", sesh.dlcData_nCams_nFrames_nImgPts_XYC,)
-            
-            sesh.dlc_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.dlcData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=.5)
[…]
