### 2021-07-26 431e360
quick fixes to get the animation up and running
