+++ freemocap/playskeleton.py
@@ -37,7 +37,6 @@ def ReplaySkeleton_matplotlib(
     useDLC=False,
 ):
 
-    startFrame = 150
     
     mean_charuco_fr_mar_dim = np.load(session.dataArrayPath/'charuco_3d_points.npy')
 
@@ -51,8 +50,9 @@ def ReplaySkeleton_matplotlib(
 
     if useDLC:
         dlc_fr_mar_dim = np.load(session.dataArrayPath / "deepLabCut_3d.npy")
-        dlcData_nCams_nFrames_nImgPts_XYC = np.load(session.dataArrayPath / "deepLabCutData_2d.npy")
-
+        dlc0 = np.squeeze(dlc_fr_mar_dim[:, 0, :])
+        dlc1 = np.squeeze(dlc_fr_mar_dim[:, 1, :])
+        dlc2 = np.squeeze(dlc_fr_mar_dim[:, 2, :])
         ballTrailLen = 4
 
     if useOpenPose:
@@ -65,8 +65,6 @@ def ReplaySkeleton_matplotlib(
         camImgPathList[cam] = list(sorted(Path(imgPathList[cam]).glob('*.png')))
         #session.numFrames = len(camImgPathList[cam]) #will need to perhaps put a check in on whether numFrames between cameras are the same
     
-    #get list of synched video paths
-    syncedVidPathList = list(sorted(session.syncedVidPath.glob('*.mp4')))
 
 # define Skeleton connections 
 # head = [17, 15, 0, 16, 18]
@@ -91,40 +89,12 @@ def ReplaySkeleton_matplotlib(
     pinky = np.array([0, 17, 18, 19, 20])
 
     fig = plt.figure(figsize=(11, 8))
-    
-    vidCapObj_list = []
-    if vidType == 0: #no videos
+
+    if vidType == 0:
         axMain = fig.add_subplot(111, projection="3d")
-    elif vidType == 1: #one video
-        numCamSubPlots = 1
+    elif vidType == 1:
         axMain = fig.add_subplot(121, projection="3d")
         axCam1 = fig.add_subplot(122)
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[1]))) #open the video file for cam 1
-    
-    elif vidType == 2: #four video
-        figGridSpec = fig.add_gridspec(2,4)
-        axMain = fig.add_subplot(figGridSpec[:2,:2],projection='3d')
-        axCam1 = fig.add_subplot(figGridSpec[0,2])
-        axCam2 = fig.add_subplot(figGridSpec[0,3])
-        axCam3 = fig.add_subplot(figGridSpec[1,2])
-        axCam4 = fig.add_subplot(figGridSpec[1,3])
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[1]))) #open the video file for cam 1
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[3]))) 
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[4]))) 
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[6]))) 
-
-    #make a list of video capture objects to load images from as needed for later plot
-    for camNum in range(numCamSubPlots):
-        vidCapObj_list = []
-        vidCapObj_list.append(cv2.VideoCapture(str(syncedVidPathList[camNum+1])))
-
-    #read over the frames preceeding startFrame, so that the video will be at the right place in time for the plotting loop
-    #(I am doing this b/c I can't figure out how to set the capture object to start reading AT startframe....)
-    # SLOPPY TECH DEBT
-    for fr in range(startFrame):
[…]
