### 2021-08-03 f87dcba
Added documentation in header and fixed some minor bugs
