### 2021-08-03 29d72b2
Added documentation in header and fixed some minor bugs
