+++ freemocap/user_preferences.yaml
@@ -1,21 +1,0 @@
-default:
-  path_to_save: C:\Users\Rontc\Documents\GitHub\freemocap
-  dlc_config_paths: []
-  rotations: {}
-  parameters:
-    exposure: -5
-    resWidth: 640
-    resHeight: 480
-    framerate: 25
-    codec: DIVX
-saved:
-  rotations:
-    1: 0
-    2: 0
-  parameters:
-    exposure: -5
-    resWidth: 640
-    resHeight: 480
-    framerate: 25
-    codec: DIVX
-  path_to_save: C:\Users\Rontc\Documents\GitHub\freemocap
