### 2021-08-04 86310c6
PYQT mediapipe fixes
