+++ freemocap/fmc_pyqtgraph.py
@@ -33,139 +33,149 @@ class PlaySkeletonWidget:
 
             self.Skel3dViewWidget.addItem(self.skelScatterItem)
 
+        elif session.useMediaPipe:
+            self.skel_fr_mar_dim = np.load(
+                session.dataArrayPath / "mediaPipeSkel_3d.npy"
+            )
+            self.skelScatterItem = gl.GLScatterPlotItem(
+                pos=self.skel_fr_mar_dim, color=(1, 0, 1, 0.8), size=10, pxMode=False
+            )
+
+            self.Skel3dViewWidget.addItem(self.skelScatterItem)
+
             # make skeleton line plots (numbers specify what order to connects the dots)
-            self.head = [17, 15, 0, 1, 0, 16, 18]
-            self.spine = [1, 8]
-            self.rArm = [17, 15, 0, 16, 18]
-            self.rArm = [4, 3, 2, 1]
-            self.lArm = [1, 5, 6, 7]
-            self.rLeg = [11, 10, 9, 8]
-            self.lLeg = [14, 13, 12, 8]
-            self.rFoot = [11, 23, 22, 11, 24]
-            self.lFoot = [14, 20, 19, 14, 21]
+        #     self.head = [17, 15, 0, 1, 0, 16, 18]
+        #     self.spine = [1, 8]
+        #     self.rArm = [17, 15, 0, 16, 18]
+        #     self.rArm = [4, 3, 2, 1]
+        #     self.lArm = [1, 5, 6, 7]
+        #     self.rLeg = [11, 10, 9, 8]
+        #     self.lLeg = [14, 13, 12, 8]
+        #     self.rFoot = [11, 23, 22, 11, 24]
+        #     self.lFoot = [14, 20, 19, 14, 21]
 
-            # Make some handy maps ;D
-            rHandIDstart = 25
-            lHandIDstart = rHandIDstart + 21
+        #     # Make some handy maps ;D
+        #     rHandIDstart = 25
+        #     lHandIDstart = rHandIDstart + 21
 
-            self.rHand = (
-                np.array(
-                    [
-                        0,
-                        1,
-                        2,
-                        3,
-                        4,
-                        False,
-                        0,
-                        5,
-                        6,
-                        7,
-                        8,
-                        False,
-                        0,
-                        9,
-                        10,
-                        11,
-                        12,
-                        False,
-                        0,
-                        13,
-                        14,
-                        15,
-                        16,
-                        False,
-                        0,
-                        17,
-                        18,
-                        19,
-                        20,
-                    ]
-                )
-                + rHandIDstart
-            )
-            self.lHand = self.rHand + lHandIDstart
+        #     self.rHand = (
+        #         np.array(
+        #             [
+        #                 0,
+        #                 1,
+        #                 2,
+        #                 3,
+        #                 4,
+        #                 False,
+        #                 0,
+        #                 5,
+        #                 6,
+        #                 7,
+        #                 8,
+        #                 False,
+        #                 0,
+        #                 9,
+        #                 10,
+        #                 11,
+        #                 12,
+        #                 False,
+        #                 0,
+        #                 13,
+        #                 14,
[…]
