### 2021-08-04 978b891
PYQT mediapipe fixes
