### 2021-08-06 0e1524a
fixing a try/except bug
