### 2021-08-06 4ec1e93
fixing a try/except bug
