### 2021-08-16 cb22641
Update READ.me with screen recording of install
I simply made a screen recording of the install process and added a link to the vimeo video in the Installation section
