+++ freemocap/fmc_openpose.py
@@ -11,7 +11,9 @@ from pathlib import Path
 
 
 def runOpenPose(session, dummyRun=False):
-
+    """
+    Run OpenPose on synced videos, and save bdoy tracking data to be parsed
+    """
     if not dummyRun:
         os.chdir(session.openPoseExePath)
 
@@ -79,84 +81,87 @@ def runOpenPose(session, dummyRun=False):
 
 # @jit(nopython=True)
 def parseOpenPose(session):
-        thisCamNum = -1
+    """
+    Parse through saved OpenPose data and create a 2D numpy array of tracked body points
+    """
+    thisCamNum = -1
 
-        ## %%
-        #build header for dataframe - NOTE - #openpose data comes in a line ordered 'pixel x location (px)', 'pixel y (py)', 'confidence (conf)' for each keypoint  
+    ## %%
+    #build header for dataframe - NOTE - #openpose data comes in a line ordered 'pixel x location (px)', 'pixel y (py)', 'confidence (conf)' for each keypoint  
+    
+    dataFrameHeader = []
+    bodyCols = 75
+    handCols = 63 #per hand
+    faceCols = 210 #das alotta face!
+    headerLength = bodyCols + 2*handCols + faceCols#should be 411 for whatever version of openpose i was using on 11 Jan 2021
+    numImgPoints = headerLength
+
+    for bb in range(0,int(bodyCols/3)): #loop through the number of body markers (i.e. #bodyCols/3)
+        dataFrameHeader.append('body_' + str(bb).zfill(3) + '_pixx')
+        dataFrameHeader.append('body_' + str(bb).zfill(3) + '_pixy')
+        dataFrameHeader.append('body_' + str(bb).zfill(3) + '_conf')
+
+    for hr in range(0,int(handCols/3)): #loop through the number of handR markers (i.e. #handCols/3)
+        dataFrameHeader.append('handR_' + str(hr).zfill(3) + '_pixx')
+        dataFrameHeader.append('handR_' + str(hr).zfill(3) + '_pixy')
+        dataFrameHeader.append('handR_' + str(hr).zfill(3) + '_conf')
+
+    for hl in range(0,int(handCols/3)): #loop through the number of handL markers (i.e. #handCols/3)
+        dataFrameHeader.append('handL_' + str(hl).zfill(3) + '_pixx')
+        dataFrameHeader.append('handL_' + str(hl).zfill(3) + '_pixy')
+        dataFrameHeader.append('handL_' + str(hl).zfill(3) + '_conf')
+
+    for ff in range(0,int(faceCols/3)): #loop through the number of Face markers (i.e. #faceCols/3)
+        dataFrameHeader.append('face_' + str(ff).zfill(3) + '_pixx')
+        dataFrameHeader.append('face_' + str(ff).zfill(3) + '_pixy')
+        dataFrameHeader.append('face_' + str(ff).zfill(3) + '_conf')
+
+    assert len(dataFrameHeader) == headerLength, ['Header is the wrong length! Should be ' +  str(headerLength) + ' but it is ' + str(len(dataFrameHeader)) + ' Check version of OpenPose?']
+
+    ## %% 
+    ## load in data from json files
+    #numFrames = int(len(list(Path(session.openPose_jsonPathList[0]).glob('*')))) #lol
+    numFrames = session.numFrames
+    numMarkers= int(int(len(dataFrameHeader)/3))
+    numCams = int(session.numCams)
+
+    openPoseData_nCams_nFrames_nImgPts_XYC = np.ndarray([numCams,numFrames,numMarkers,3]) #hardcoding for now because I am a bad person
+
[…]
