### 2021-08-27 c458f2e
Generalizing skeleton animation
Dev Update 2020-08-27 https://twitter.com/freemocap/status/1431235339260157959
