+++ freemocap/__init__.py
@@ -3,13 +3,38 @@ from ._version import get_versions
 __version__ = get_versions()["version"]
 del get_versions
 
-from freemocap import fmc_runme, fmc_demo
+from freemocap.fmc_runme import RunMe
 
-def RunMe(sessionID=None,stage=1,useOpenPose=True, openPoseDummyRun = False, useMediaPipe=False,useDLC=False,dlcConfigPath=None,debug=False,setDataPath=False,userDataPath = None, recordVid=False):
-    """ 
-    Passes user-set values from the freemocap_runme_script to the fmc_runme.Run function (and allows the user to run the pipeline with just an fmc.RunMe command)
-    """  
-    fmc_runme.Run(sessionID,stage,useOpenPose,openPoseDummyRun, useMediaPipe,useDLC,dlcConfigPath,debug,setDataPath,userDataPath,recordVid)
+# from freemocap import fmc_runme, fmc_demo #JSM NOTE - there's gotta be a better way to do this lol
+
+# def RunMe(sessionID=None,
+#             stage=1,
+#             useOpenPose=True, 
+#             openPoseDummyRun=False,
+#             useMediaPipe=False,
+#             useDLC=False,
+#             dlcConfigPath=None,
+#             debug=False,
+#             setDataPath=False,
+#             userDataPath = None, 
+#             recordVid=False,
+#             reconstructionConfidenceThreshold = .8,
+#             charucoSquareSize = 36,
+#             ):
+#     """ 
+#     Passes user-set values from the freemocap_runme_script to the fmc_runme.Run function (and allows the user to run the pipeline with just an fmc.RunMe command)
+#     """  
+#     fmc_runme.Run(sessionID,
+#     stage,
+#     useOpenPose,
+#     openPoseDummyRun, 
+#     useMediaPipe,
+#     useDLC,
+#     dlcConfigPath,
+#     debug,
+#     setDataPath,
+#     userDataPath,
+#     recordVid)
 
 def RunDemo():
     """ 
