+++ freemocap/play_skeleton_animation.py
@@ -1,4 +1,3 @@
-
 """
 ============
 ██████  ██       █████  ██    ██     ███████ ██   ██ ███████ ██      ███████ ████████  ██████  ███    ██      █████  ███    ██ ██ ███    ███  █████  ████████ ██  ██████  ███    ██ 
@@ -38,6 +37,7 @@ from matplotlib.widgets import Slider
 import cv2 
 from scipy.signal import savgol_filter
 
+
 import copy
 from pathlib import Path
 
@@ -63,21 +63,19 @@ def PlaySkeletonAnimation(
     startFrame=0,
     azimuth=-90,
     elevation=-61,
-    numCams=4,
-    useCams = None,
     useOpenPose=True,
     useMediaPipe=False,
     useDLC=False,
     recordVid = True
     ):
-###  
-###
-###      ██    ██ ██████  ██████   █████  ████████ ███████         ███████ ██  ██████  ██    ██ ██████  ███████ 
-###      ██    ██ ██   ██ ██   ██ ██   ██    ██    ██              ██      ██ ██       ██    ██ ██   ██ ██      
-###      ██    ██ ██████  ██   ██ ███████    ██    █████           █████   ██ ██   ███ ██    ██ ██████  █████   
-###      ██    ██ ██      ██   ██ ██   ██    ██    ██              ██      ██ ██    ██ ██    ██ ██   ██ ██      
-###       ██████  ██      ██████  ██   ██    ██    ███████         ██      ██  ██████   ██████  ██   ██ ███████ 
-###                                                                                                               
+    ###  
+    ###
+    ###      ██    ██ ██████  ██████   █████  ████████ ███████         ███████ ██  ██████  ██    ██ ██████  ███████ 
+    ###      ██    ██ ██   ██ ██   ██ ██   ██    ██    ██              ██      ██ ██       ██    ██ ██   ██ ██      
+    ###      ██    ██ ██████  ██   ██ ███████    ██    █████           █████   ██ ██   ███ ██    ██ ██████  █████   
+    ###      ██    ██ ██      ██   ██ ██   ██    ██    ██              ██      ██ ██    ██ ██    ██ ██   ██ ██      
+    ###       ██████  ██      ██████  ██   ██    ██    ███████         ██      ██  ██████   ██████  ██   ██ ███████ 
+    ###                                                                                                               
     def update_figure(frameNum):
         """ 
         Called by matplotlib animator for each frame.
@@ -85,6 +83,9 @@ def PlaySkeletonAnimation(
         
         skel_dottos = matplotlib_artist_objs['skel_dottos'] 
         skel_trajectories = figure_data['skel_trajectories|mar|fr_dim']
+
+        charuco_dottos = matplotlib_artist_objs['charuco_dottos'] 
+        charuco_trajectories = figure_data['charuco_trajectories|mar|fr_dim']
 
         if useDLC:
             dlc_dottos = matplotlib_artist_objs['dlc_dottos'] 
@@ -113,20 +114,27 @@ def PlaySkeletonAnimation(
         update_3d_skel_segments('lHand', skel_fr_mar_dim)
         update_3d_skel_segments('face', skel_fr_mar_dim)
 
-        update_3d_skel_segments('pinkBall', dlc_fr_mar_dim)
-        update_3d_skel_segments('redBall', dlc_fr_mar_dim)
-        update_3d_skel_segments('greenBall', dlc_fr_mar_dim)
-        update_3d_skel_segments('wobbleBoard', dlc_fr_mar_dim)
[…]
