+++ freemocap/fmc_runme.py
@@ -25,7 +25,7 @@ import numpy as np
 
 from ruamel.yaml import YAML
 
-def Run(sessionID=None,
+def RunMe(sessionID=None,
         stage=1,
         useOpenPose=True, 
         openPoseDummyRun = False, 
@@ -35,7 +35,11 @@ def Run(sessionID=None,
         debug=False,
         setDataPath = False,
         userDataPath = None,
-        recordVid = True,):
+        recordVid = True,
+        reconstructionConfidenceThreshold = .8,
+        charucoSquareSize = 36,#mm - ~the size of the squares when printed on 8.5x11" paper based on parameters in ReadMe.md
+        calVideoFrameLength = 120,
+        ):
     """ 
     Starts the freemocap pipeline based on either user-input values, or default values. Creates a new session class instance (called sesh)
     based on the specified inputs. Checks for previous user preferences and choices if they exist, or will prompt the user for new choices
@@ -157,8 +161,8 @@ def Run(sessionID=None,
                         #marker_length=.8,
                         #  square_length = 121, #big boi charuco
                         #  marker_length = 98,
-                        square_length = 82,#regular boi charuco
-                        marker_length = 65,
+                        square_length = charucoSquareSize,#mm
+                        marker_length = charucoSquareSize*.8,#mm
                         marker_bits=4, dict_size=250)
 
 
@@ -197,23 +201,25 @@ def Run(sessionID=None,
     if stage <= 3:
         print()
         print('Starting Calibration')
-        sesh.cgroup, sesh.mean_charuco_fr_mar_dim = calibrate.CalibrateCaptureVolume(sesh,board)
+        sesh.cgroup, sesh.mean_charuco_fr_mar_dim = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength)
     else:
         print('Skipping Calibration')
 
     # %% Stage Four
     if stage <= 4:
+
         print('Starting Track Image Points')
         if sesh.useMediaPipe:
             fmc_mediapipe.runMediaPipe(sesh)
             sesh.mediaPipeData_nCams_nFrames_nImgPts_XYC = fmc_mediapipe.parseMediaPipe(sesh)
-            sesh.mediaPipeSkel_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.mediaPipeData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=.5)
+            sesh.mediaPipeSkel_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.mediaPipeData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=reconstructionConfidenceThreshold)
             np.save(sesh.dataArrayPath/'mediaPipeSkel_3d.npy', sesh.mediaPipeSkel_fr_mar_dim) #save data to npy
         sesh.save_session()
+
         if sesh.useOpenPose:
             fmc_openpose.runOpenPose(sesh, dummyRun=openPoseDummyRun)
             sesh.openPoseData_nCams_nFrames_nImgPts_XYC = fmc_openpose.parseOpenPose(sesh)
-            sesh.openPoseSkel_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.openPoseData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=.5)
+            sesh.openPoseSkel_fr_mar_dim = reconstruct3D.reconstruct3D(sesh,sesh.openPoseData_nCams_nFrames_nImgPts_XYC, confidenceThreshold=reconstructionConfidenceThreshold)
[…]
