+++ freemocap/webcam/recordGUI.py
@@ -17,6 +17,7 @@ class recordGUI:
     # a listbox, and allows you to pick a task for the cameras
     def __init__(self, master):
         self.master = master
+        self.refresh = False 
         # Detect how many inputs you have available
         available_inputs = checkcams.CreateAvailableCamList()
         number_of_cams = len(available_inputs)
@@ -61,12 +62,21 @@ class recordGUI:
         self.sub_button = Button(bottom_frame, text="Submit", command=self.Submit)
         self.sub_button.pack(side=tk.RIGHT)
 
+        self.refresh_button = Button(bottom_frame, text="Refresh", command=self.Refresh)
+        self.refresh_button.pack(side=tk.RIGHT)
+
     def Submit(self):
         self.taskChoice = self.currentTask.get()
         self.selected = [
             int(self.list_box.get(pos)) for pos in self.list_box.curselection()
         ]
+        self.refresh = False
         self.master.destroy()
+    
+    def Refresh(self):
+        self.refresh = True
+        self.master.destroy()
+
 
 
 class SettingsGUI:
@@ -245,6 +255,7 @@ class SettingsGUI:
 class ProceedToRecordGUI:
     def __init__(self, master, sessionID_in, current_path_to_save):
         self.master = master
+        self.restart_setup = False
         self.sessionID_in = sessionID_in
         self.continueToRecording = ""
         self.sessionID_out = ""
@@ -274,8 +285,10 @@ class ProceedToRecordGUI:
 
             # master.title("Proceed to Recording?")
         self.proceed_button = Button(text="Proceed", command=self.proceed)
+        self.change_button = Button(text = "Change Parameters", command = self.change_params)
         self.stop_button = Button(text="Quit", command=self.stop)
         self.stop_button.pack(side=tk.RIGHT)
+        self.change_button.pack(side = tk.LEFT)
         self.proceed_button.pack(side=tk.RIGHT)
 
 
@@ -289,6 +302,13 @@ class ProceedToRecordGUI:
     def stop(self):
         self.master.destroy()
 
+    def change_params(self):
+        self.restart_setup = True
+        self.sessionID_out = self.sessionIDEntry.get()
+        self.save_path = self.currentPath
+        self.master.destroy()
+
+
     def proceed(self):
         self.continueToRecording = True
         self.sessionID_out = self.sessionIDEntry.get()
@@ -296,18 +316,26 @@ class ProceedToRecordGUI:
         self.master.destroy()
 
 
-def RunGUI(sessionID_in, rotation_entry, parameter_entry,current_path_to_save):
+
+def RunChoiceGUI():
 
     # ---Get the camera inputs and the task
-    root = tk.Tk()
-    camera_choice = recordGUI(root)
-    root.mainloop()
+    refresh = True
+    while refresh:            
+        root = tk.Tk()
+        camera_choice = recordGUI(root)
+        root.mainloop()
+        refresh = camera_choice.refresh
 
     cam_inputs = camera_choice.selected
     if not cam_inputs:
         raise ValueError("No camera inputs selected")
     task = camera_choice.taskChoice
 
+    return cam_inputs,task
+
+
+def RunParametersGUI(sessionID_in, rotation_entry, parameter_entry,current_path_to_save,cam_inputs,task):
     # ---Get all the necessary parameters
     root = tk.Tk()
     recording_settings = SettingsGUI(
@@ -327,7 +355,7 @@ def RunGUI(sessionID_in, rotation_entry, parameter_entry,current_path_to_save):
         sessionID = None
         savepath = None
 
-    return task, cam_inputs, recording_settings.rotation_output, paramDict, sessionID,savepath
[…]
