### 2021-08-28 302e26b
altered calibration method to produce plottable charuco tracks
