### 2021-08-28 82cae58
altered calibration method to produce plottable charuco tracks
