+++ freemocap/fmc_pyqtgraph.py
@@ -24,21 +24,21 @@ class PlaySkeletonWidget:
         # load data to be plotted
         self.useOpenPose = session.useOpenPose
         if self.useOpenPose:
-            self.skel_fr_mar_dim = np.load(
+            self.skel_fr_mar_xyz = np.load(
                 session.dataArrayPath / "openPoseSkel_3d.npy"
             )
             self.skelScatterItem = gl.GLScatterPlotItem(
-                pos=self.skel_fr_mar_dim, color=(1, 0, 1, 0.8), size=10, pxMode=False
+                pos=self.skel_fr_mar_xyz, color=(1, 0, 1, 0.8), size=10, pxMode=False
             )
 
             self.Skel3dViewWidget.addItem(self.skelScatterItem)
 
         elif session.useMediaPipe:
-            self.skel_fr_mar_dim = np.load(
+            self.skel_fr_mar_xyz = np.load(
                 session.dataArrayPath / "mediaPipeSkel_3d.npy"
             )
             self.skelScatterItem = gl.GLScatterPlotItem(
-                pos=self.skel_fr_mar_dim, color=(1, 0, 1, 0.8), size=10, pxMode=False
+                pos=self.skel_fr_mar_xyz, color=(1, 0, 1, 0.8), size=10, pxMode=False
             )
 
             self.Skel3dViewWidget.addItem(self.skelScatterItem)
@@ -97,7 +97,7 @@ class PlaySkeletonWidget:
         #     self.lHand = self.rHand + lHandIDstart
 
         #     self.headLineItem = gl.GLLinePlotItem(
-        #         pos=self.skel_fr_mar_dim[:, self.head, :],
+        #         pos=self.skel_fr_mar_xyz[:, self.head, :],
         #         color=(1, 1, 1, 0.2),
         #         width=4.0,
         #         antialias=True,
@@ -105,7 +105,7 @@ class PlaySkeletonWidget:
         #     self.Skel3dViewWidget.addItem(self.headLineItem)
 
         #     self.spineLineItem = gl.GLLinePlotItem(
-        #         pos=self.skel_fr_mar_dim[:, self.spine, :],
+        #         pos=self.skel_fr_mar_xyz[:, self.spine, :],
         #         color=(1, 1, 1, 0.9),
         #         width=4.0,
         #         antialias=True,
@@ -113,7 +113,7 @@ class PlaySkeletonWidget:
         #     self.Skel3dViewWidget.addItem(self.spineLineItem)
 
         #     self.rArmLineItem = gl.GLLinePlotItem(
-        #         pos=self.skel_fr_mar_dim[:, self.rArm, :],
+        #         pos=self.skel_fr_mar_xyz[:, self.rArm, :],
         #         color=(1, 0.2, 0.15, 0.9),
         #         width=4.0,
         #         antialias=True,
@@ -121,7 +121,7 @@ class PlaySkeletonWidget:
         #     self.Skel3dViewWidget.addItem(self.rArmLineItem)
 
         #     self.lArmLineItem = gl.GLLinePlotItem(
-        #         pos=self.skel_fr_mar_dim[:, self.lArm, :],
+        #         pos=self.skel_fr_mar_xyz[:, self.lArm, :],
         #         color=(0.4, 0.6, 1, 0.9),
         #         width=4.0,
         #         antialias=True,
@@ -129,7 +129,7 @@ class PlaySkeletonWidget:
         #     self.Skel3dViewWidget.addItem(self.lArmLineItem)
 
         #     self.rLegLineItem = gl.GLLinePlotItem(
-        #         pos=self.skel_fr_mar_dim[:, self.rLeg, :],
+        #         pos=self.skel_fr_mar_xyz[:, self.rLeg, :],
[…]
