+++ freemocap/play_skeleton_animation.py
@@ -81,8 +81,13 @@ def PlaySkeletonAnimation(
         Called by matplotlib animator for each frame.
         """
         
-        skel_dottos = matplotlib_artist_objs['skel_dottos'] 
-        skel_trajectories = figure_data['skel_trajectories|mar|fr_dim']
+        if useOpenPose:
+            op_skel_dottos = matplotlib_artist_objs['op_skel_dottos'] 
+            op_skel_trajectories = figure_data['openPose_skel_trajectories|mar|fr_dim']
+
+        if useMediaPipe:
+            mp_skel_dottos = matplotlib_artist_objs['mp_skel_dottos'] 
+            mp_skel_trajectories = figure_data['mediaPipe_skel_trajectories|mar|fr_dim']
 
         charuco_dottos = matplotlib_artist_objs['charuco_dottos'] 
         charuco_trajectories = figure_data['charuco_trajectories|mar|fr_dim']
@@ -92,34 +97,48 @@ def PlaySkeletonAnimation(
             dlc_trajectories = figure_data['dlc_trajectories|mar|fr_dim']
 
         #function to update the lines for each 3d body segment
-        def update_3d_skel_segments(key, data_fr_mar_dim):       
+        def update_3d_skel_segments(key, data_fr_mar_xyz):       
             """ 
             updates the Artist of each body segment with the current frame data.
             """    
-            try:   
+            split = key.split('_')
+
+            if split[0] == 'op':
                 dict_of_segments_idxs = dict_of_openPoseSegmentIdx_dicts[key]
-            except: 
+            if split[0] == 'mp':
+                dict_of_segments_idxs = dict_of_mediaPipeSegmentIdx_dicts[key]
+            elif split[0] == 'dlc':
                 dict_of_segments_idxs = dict_of_dlcSegmentIdx_dicts[key]
 
-            dict_of_artists = matplotlib_artist_objs[key] 
+            try:
+                dict_of_artists = matplotlib_artist_objs[key] 
+            except:
+                print('KeyError:', key)
+                
 
             for thisItem in dict_of_segments_idxs.items():
                 segName = thisItem[0]
                 segArtist = dict_of_artists[segName]
-                segArtist.set_data((data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 0], data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 1]))
-                segArtist.set_3d_properties(data_fr_mar_dim[frameNum, dict_of_segments_idxs[segName], 2])                
-            
-        update_3d_skel_segments('body', skel_fr_mar_dim)
-        update_3d_skel_segments('rHand', skel_fr_mar_dim)
-        update_3d_skel_segments('lHand', skel_fr_mar_dim)
-        update_3d_skel_segments('face', skel_fr_mar_dim)
+                segArtist.set_data((data_fr_mar_xyz[frameNum, dict_of_segments_idxs[segName], 0], data_fr_mar_xyz[frameNum, dict_of_segments_idxs[segName], 1]))
+                segArtist.set_3d_properties(data_fr_mar_xyz[frameNum, dict_of_segments_idxs[segName], 2])                
+
+        if useOpenPose:    
+            update_3d_skel_segments('op_body', openPose_skel_fr_mar_xyz)
+            update_3d_skel_segments('op_rHand', openPose_skel_fr_mar_xyz)
[…]
