+++ freemocap/user_preferences.yaml
@@ -10,3 +10,13 @@ default:
     codec: DIVX
 saved:
   path_to_save: C:/Users/jonma/Dropbox/FreeMoCapProject
+  rotations:
+    0: 0
+    2: 0
+    3: 0
+  parameters:
+    exposure: -7
+    resWidth: 640
+    resHeight: 480
+    framerate: 30
+    codec: DIVX
