### 2021-08-30 9c1cfe6
Tell `doctr` there is no `master`
https://drdoctr.github.io/recipes.html#deploy-docs-from-a-non-master-branch
