### 2021-08-30 1bae3fd
joint openpose and mediapipe reconstructions more or less working :D
Needs - 

- Method to spit out charuco dots overlaid on videos, to help diagnose Anipose failings

- method to adjust webcam parameters before recording, e.g. the 'opencv' example here - 
https://realpython.com/pysimplegui-python/
