+++ freemocap_runme_script.py
@@ -1,18 +1,18 @@
 from freemocap.fmc_mediapipe import runMediaPipe
 import freemocap as fmc
 from pathlib import Path
-import socket 
+import sys
 
-try:
-    fmc.RunMe(  stage=5,
-                useOpenPose=False, 
-                runOpenPose=True, 
-                useMediaPipe=True, 
-                runMediaPipe = True,
-                useDLC=False, 
-                recordVid = True,
-                reconstructionConfidenceThreshold = 0.5,
-                calVideoFrameLength = -1) 
-except:
-    print("Unexpected error:", sys.exc_info()[0])
-    raise
+
+fmc.RunMe(
+    sessionID = 'sesh_2021-08-29_15_26_34',
+    stage = 4,
+    useOpenPose=True, 
+    runOpenPose=True, 
+    useMediaPipe=True, 
+    runMediaPipe = True,
+    useDLC=False, 
+    recordVid = True,
+    reconstructionConfidenceThreshold = 0.5,
+    calVideoFrameLength = -1,
+)
