### 2021-08-30 6a9175f
Revert "Merge pull request #49 from jonmatthis/feature/update_mediapipe"
This reverts commit cb03aa345862ce6658b83bde1788fc11cff85295, reversing
changes made to 1bae3fd4a06bfb135f82c67a18cd53d33e613b6e.
