### 2021-09-01 bb8c751
Revert "Fixing ImportError: Cannot load backend 'WXAgg'"
Installing `xvfb` causes Travis to hand for 10 minutes, after
which it fails by timeout.

This reverts commit df74d3fe912e1de8bf48c6574dce9d3b63c7fe3e.
