### 2021-09-01 827b112
.travis.yml: Run `pytest` in verbose mode to debug freeze
`No output has been received in the last 10m0s.` is how the test are
currently failing, which started after installing `xvfb` in df74d3f . If
the attempt to thaw it out won't succeeded, the commit would need
to be reverted, because partial test result are better that no tests.
