### 2021-09-01 01edd55
Try --verbose for pytest in another PR
