+++ .travis.yml
@@ -21,7 +21,7 @@ install:
   - pip install -r requirements-dev.txt
 
 script:
-  - coverage run -m pytest  # Run the tests and check for test coverage.
+  - coverage run -m pytest -v # Run the tests and check for test coverage.
   - coverage report -m  # Generate test coverage report.
   - codecov  # Upload the report to codecov.
   #- flake8  # Enforce code style ('relaxed' line length limit is set in .flake8 config file).
