### 2021-09-01 #41 @lwasser
hey @jonmatthis checking back in here on this. where can  we start with helping you? could you identify perhaps  one area to start with so i can find people to provide support? i can also answer questions if t hat is helpful or we can chat if that  would be more helpful.
