### 2021-09-02 #41 @jonmatthis
Hello! Thanks for checking back in!

This repo has garnered quite a bit of interest since my original post, so I am currently working on getting the core functionality to a usable 'alpha-ish' state, before coming back to clean up the inner structural and package aspects of the project in hopes of publishish with PyOpenSci and/or JOSS within the next few months (and start looking for funding sources to support it on a similar time scale).

Per your above requests for more information, probably the easiest thing would be to just direct y'all to the website - freemocap.org or more specifcally the white-paper-like abstract I wrote for it a while ago (maybe slightly out of date, but still generall On-Vibe) - http://jonmatthis.com/assets/fmc-whitepaper.pdf

I will DEFINITELY need help cleaning this package up before submission to PyOpenSci and JOSS, but at the moment I think the most immediate help I need is with cleaning up the pip install and dependencies of this thing.

Specifically, we recently managed to come up with a configuration of this software that works without the need for a Cuda GPU, which I think is very important for ensuring broader adoption and impact.

So I would really like to set up the `pip install` to allow for multiple configurations of this software that just installs the 'base' version (which doesn't need an Nvidia CUDA GPU) as well as a 'full' version that includes support for CUDA-enabled packages like [DeepLabCut](http://www.mackenziemathislab.org/deeplabcut) and [OpenPose](https://github.com/CMU-Perceptual-Computing-Lab/openpose)

Ideally, the use could pip install in the following ways -

`pip install freemocap` <- installs the base version, without CUDA stuff
`pip install freemocap[cuda]` which installs the base version, plus additional CUDA-requiring packages (e.g. tensorflow, deeplabcut, etc)

It would also be great if I could redirect different pip install specifications to the same install, e.g. so `pip install freemocap[cuda]`, `pip install freemocap[full]` and`pip install freemocap[deeplabcut]` all install the same "Full" version of this package.

Any assistance would be great! Even if you're just pointing me towards websites/tutorials/etc that explain how to do the things (or even just giving me the google-able terms to help me find those resources myself!)

thanks!

(also, PS - The most up-to-date version of this software lives on the `Develop` branch of github.com/jonmatthis/freemocap )
