### 2021-09-07 4c6b7ce
fix looping GUI bug when selecting 'record'
