+++ freemocap/fmc_anipose.py
@@ -1694,11 +1694,11 @@ class CameraGroup:
                 pprint(get_connections(rtvecs, self.get_names()))
 
                 
-                for camNum in range(rtvecs.shape[0]):
-                    if np.sum(np.isnan(rtvecs[camNum,:,:])) == rtvecs[camNum,:,:].size:
-                        rtvecsOG = rtvecs.copy() #for posterity
-                        rtvecs = np.delete(rtvecs, 0,0)
-                        print('REMOVING CAM# {} BECAUSE ANIPOSE HAD IT AS NaNs FOR SOME UNKNOWN REASON'.format(camNum))
+                # for camNum in range(rtvecs.shape[0]):
+                #     if np.sum(np.isnan(rtvecs[camNum,:,:])) == rtvecs[camNum,:,:].size:
+                #         rtvecsOG = rtvecs.copy() #for posterity
+                #         rtvecs = np.delete(rtvecs, 0,0)
+                #         print('REMOVING CAM# {} BECAUSE ANIPOSE HAD IT AS NaNs FOR SOME UNKNOWN REASON'.format(camNum))
                         
                 rvecs, tvecs = get_initial_extrinsics(rtvecs)
                 self.set_rotations(rvecs)
