+++ freemocap/analysis/skel_analisys.py
@@ -1,4 +1,6 @@
 #%%
+%matplotlib
+
 import numpy as np
 
 import matplotlib.pyplot as plt
@@ -21,37 +23,53 @@ from rich.markdown import Markdown
 
 # %% load dlc data 
 
-openPose3dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-08-27_17_31_30\DataArrays\openPoseSkel_3d.npy")
+openPose3dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-09-07_09_09_56\DataArrays\openPoseSkel_3d.npy")
 
-mediaPipe3dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-08-27_17_31_30\DataArrays\mediaPipeSkel_3d.npy")
+mediaPipe3dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-09-07_09_09_56\DataArrays\mediaPipeSkel_3d.npy")
+mediaPipe3dReprojPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-09-07_09_09_56\DataArrays\mediaPipeSkel_reprojErr.npy")
 
-openPose2dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-08-27_17_31_30\DataArrays\openPoseData_2d.npy")
+openPose2dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-09-07_09_09_56\DataArrays\openPoseData_2d.npy")
 
-mediaPipe2dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-08-27_17_31_30\DataArrays\mediaPipe_2d.npy")
+mediaPipe2dDataPath = Path(r"C:\Users\jonma\Dropbox\FreeMoCapProject\FreeMocap_Data\sesh_2021-09-07_09_09_56\DataArrays\mediaPipe_2d.npy")
 
-#%%
+#%% pull out 2d data
 mediaPipe_nCams_nFrames_nImgPts_XYC = np.load(mediaPipe2dDataPath)
 
 mp_rToe_conf = mediaPipe_nCams_nFrames_nImgPts_XYC[:,:,32,2].transpose()
 mp_nose_conf = mediaPipe_nCams_nFrames_nImgPts_XYC[:,:,0,2].transpose()
 
+#%% pull out 3d reproj error
+
+mediaPipe_ReProjErr= np.load(mediaPipe3dReprojPath)
+
+mediaPipe_meanReproj_fr = np.nanmean(mediaPipe_ReProjErr,0)
+
 
 # %% hand/ball vs time
-fig1 = plt.figure()
-axToe = fig1.add_subplot(1,2,1)
 
-axToe.plot(mp_rToe_conf, label='Right Toe')
+camLabels = ["Cam#{}".format(str(camNum)) for camNum in range(mediaPipe_nCams_nFrames_nImgPts_XYC.shape[0])]
+
+fig1 = plt.figure()
+axToe = fig1.add_subplot(1,3,1)
+
+axToe.plot(mp_rToe_conf, label='rToe')
 axToe.set_xlabel('Time (frames)')
-axToe.set_ylabel('Confidence')
+axToe.set_ylabel('Nose Confidence')
[…]
