+++ freemocap_runme_script.py
@@ -3,16 +3,16 @@ import freemocap as fmc
 from pathlib import Path
 import sys
 
-
 fmc.RunMe(
-    sessionID = 'sesh_2021-08-29_15_26_34',
-    stage = 4,
+    # sessionID = 'sesh_2021-08-29_17_55_45',
+    sessionID=None,
+    stage = 5,
     useOpenPose=True, 
     runOpenPose=True, 
-    useMediaPipe=True, 
+    useMediaPipe=False,
     runMediaPipe = True,
     useDLC=False, 
-    recordVid = True,
-    reconstructionConfidenceThreshold = 0.5,
-    calVideoFrameLength = -1,
+    recordVid = True ,
+    reconstructionConfidenceThreshold = 0.0,
+    calVideoFrameLength = 60,
 )
