+++ freemocap/user_preferences.yaml
@@ -12,11 +12,13 @@ saved:
   path_to_save: C:/Users/jonma/Dropbox/FreeMoCapProject
   rotations:
     0: 0
+    1: 0
     2: 0
     3: 0
+    4: 0
   parameters:
     exposure: -7
-    resWidth: 960
-    resHeight: 720
+    resWidth: 640
+    resHeight: 480
     framerate: 30
     codec: DIVX
