### 2021-09-10 a8ef53d
startup and initialization reorganization
