### 2021-09-10 df9f333
startup and initialization reorganization
