+++ freemocap/fmc_runme.py
@@ -2,7 +2,6 @@ from freemocap import (
     createvideo,
     recordingconfig,
     runcams,
-    runmeGUI,
     calibrate,
     fmc_mediapipe,
     fmc_openpose,
@@ -11,10 +10,12 @@ from freemocap import (
     playskeleton,
     play_skeleton_animation,
     session,
-    webcamGUI
 )
 
 from freemocap.fmc_pyqtgraph import PlayerDockedWindow
+
+from freemocap.fmc_startup import startup
+from freemocap.webcam import camera_settings, timesync
 
 from pathlib import Path
 import os
@@ -57,106 +58,22 @@ def RunMe(sessionID=None,
     sesh.userDataPath = userDataPath
     sesh.dataFolderName = recordingconfig.dataFolder
 
-    if sesh.useDLC and stage < 5: 
-        import deeplabcut as dlc
-    #    sesh.dlcConfigPath = dlcConfigPath
+    # %% Startup 
+    startup.get_user_preferences(sesh,stage)
 
+    if sesh.useDLC and stage<5:
+         import deeplabcut as dlc
+         dlc_config_paths = startup.get_dlc_paths(session)
 
-
-    #%% load user preferences if they exist, create a new preferences yaml if they don't
-    here = Path(__file__).parent
-    preferences_path = here/'user_preferences.yaml'
-    preferences_yaml = YAML()
-
-    if preferences_path.exists():
-        preferences = preferences_yaml.load(preferences_path)
-    else:
-        preferences = recordingconfig.parameters_for_yaml
-        preferences_yaml.dump(preferences, preferences_path)
-    
-    sesh.preferences = preferences
-    sesh.preferences_path = preferences_path
-        
-
-    if sesh.useDLC and stage < 5:
-
-        try:
-            saved_dlc_paths = preferences['saved']['dlc_config_paths']
-        except: 
-            saved_dlc_paths = preferences['default']['dlc_config_paths']
-
-
-        dlc_config_paths = runmeGUI.RunChooseDLCPathGUI(sesh,saved_dlc_paths)
-        
-
-        sesh.preferences['saved']['dlc_config_paths'] = dlc_config_paths
-        sesh.save_user_preferences(sesh.preferences)
-
-        # sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Dropbox\\GitKrakenRepos\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml")
-        #sesh.dlcConfigPath = Path("C:\\Users\\jonma\\Desktop\\freemocap\\DLC_Models\\PinkGreenRedJugglingBalls-JSM-2021-05-31\\config.yaml") 
     if stage > 1:
-        #if we are rerunning a session folder
-        # 1) Check if we're using the last saved dataFolderPath, or if the user wants to choose a different one
-        #   a. if the user wants to choose one, bring up a GUI to let them decide
-        #   b. if we're using the last known path - parse the user preferences yaml (and check if that yaml exists)
-        # 2) Check that the data folder exists
-        # 3) If no sessionID was user-input, search the chosen directory for the last session created
-        if sesh.setDataPath == True:
-            sesh.basePath = runmeGUI.RunChooseDataPathGUI(session)
-            sesh.basePath = Path(sesh.basePath)
-            #sesh.dataFolderPath = Path(basePath)/sesh.dataFolderName
-
-        elif sesh.userDataPath is not None:
-            sesh.basePath = sesh.userDataPath
-        else:
-            try:
-                current_path_to_data = preferences['saved']['path_to_save']
-                sesh.basePath = current_path_to_data
-            except KeyError:
-                print('Saved Data path not found, please choose a new one')
-                sesh.basePath = runmeGUI.RunChooseDataPathGUI(session)
[…]
