### 2021-09-11 7f9dfae
fix license attribute to (correctly) read AGPLv3
