### 2021-09-12 14c347a
Cleaning up repo
A lot of the deleted files were generated by the cookiecutter I used a while back. I'm not sure what many of them do, so I am removing  them for now. Hopefully it doesn't break everything lol
