+++ freemocap_runme_script.py
@@ -4,15 +4,15 @@ from pathlib import Path
 import sys
 
 fmc.RunMe(
-    # sessionID = 'sesh_2021-08-29_17_55_45',
-    sessionID=None,
+    sessionID = 'sesh_2021-09-09_17_23_44',
+    # sessionID=None,
     stage = 5,
-    useOpenPose=True, 
+    useOpenPose=False, 
     runOpenPose=True, 
-    useMediaPipe=False,
+    useMediaPipe=True,
     runMediaPipe = True,
     useDLC=False, 
     recordVid = True ,
-    reconstructionConfidenceThreshold = 0.0,
+    reconstructionConfidenceThreshold = 0.5,
     calVideoFrameLength = 60,
 )
