### 2021-09-12 67f4f2e
Removed ffmpeg reliance from animation writer
New method is kinda dumb, but it works - Start by saving to a `gif`, then convert the gif to an `mp4` (and delete the gif)
