### 2021-09-12 d990c78
Swapping `CAP_DSHOW` with `CAP_ANY` for Ubuntu (and mac?) support
Based on suggestion from https://github.com/jonmatthis/freemocap/issues/52
