### 2021-09-12 4673df1
Swapping `CAP_DSHOW` with `CAP_ANY` for Ubuntu (and mac?) support
Based on suggestion from https://github.com/jonmatthis/freemocap/issues/52
