+++ freemocap/setup.py
@@ -1,72 +1,0 @@
-from os import path
-from setuptools import setup, find_packages
-import sys
-import versioneer
-
-
-# NOTE: This file must remain Python 2 compatible for the foreseeable future,
-# to ensure that we error out properly for people with outdated setuptools
-# and/or pip.
-min_version = (3, 7)
-if sys.version_info < min_version:
-    error = """
-freemocap does not support Python {0}.{1}.
-Python {2}.{3} and above is required. Check your Python version like so:
-
-python3 --version
-
-This may be due to an out-of-date pip. Make sure you have pip >= 9.0.1.
-Upgrade pip like so:
-
-pip install --upgrade pip
-""".format(
-        *(sys.version_info[:2] + min_version)
-    )
-    sys.exit(error)
-
-here = path.abspath(path.dirname(__file__))
-
-with open(path.join(here, "README.md"), encoding="utf-8") as readme_file:
-    readme = readme_file.read()
-
-with open(path.join(here, "requirements.txt")) as requirements_file:
-    # Parse requirements.txt, ignoring any commented-out lines.
-    requirements = [
-        line
-        for line in requirements_file.read().splitlines()
-        if not line.startswith("#")
-    ]
-
-
-setup(
-    name="freemocap",
-    version=versioneer.get_version(),
-    cmdclass=versioneer.get_cmdclass(),
-    description="The FreeMoCap Project: A free-and-open-source, hardware-and-software-agnostic, research-grade, minimal-cost motion capture system and platform for decentralized scientific research, education, and training",
-    long_description=readme,
-    author="Jonathan Samir Matthis",
-    author_email="jonmatthis@gmail.com",
-    url="https://github.com/jonmatthis/freemocap",
-    python_requires=">={}".format(".".join(str(n) for n in min_version)),
-    packages=find_packages(exclude=["docs", "tests"]),
-    entry_points={
-        "console_scripts": [
-            # 'command = some.module:some_function',
-        ],
-    },
-    include_package_data=True,
-    package_data={
-        "freemocap": [
-            # When adding files here, remember to update MANIFEST.in as well,
-            # or else they will not be included in the distribution on PyPI!
-            # 'path/to/data_file',
-        ]
-    },
-    install_requires=requirements,
-    license="AGPLv3",
-    classifiers=[
-        "Development Status :: 2 - Pre-Alpha",
-        "Natural Language :: English",
-        "Programming Language :: Python :: 3",
-    ],
-)
