+++ docs/source/installation.rst
@@ -1,0 +1,7 @@
+============
+Installation
+============
+
+At the command line::
+
+    $ pip install freemocap
