+++ README.rst
@@ -1,0 +1,20 @@
+=================
+Your Project Name
+=================
+
+.. image:: https://img.shields.io/travis//your-project-name.svg
+        :target: https://travis-ci.org//your-project-name
+
+.. image:: https://img.shields.io/pypi/v/your-project-name.svg
+        :target: https://pypi.python.org/pypi/your-project-name
+
+
+Python package for doing science.
+
+* Free software: Affero GNU Public Licence (AGPLv3)
+* Documentation: (COMING SOON!) https://.github.io/your-project-name.
+
+Features
+--------
+
+* TODO
