+++ requirements.txt
@@ -1,11 +1,21 @@
 aniposelib~=0.4.1
 h5py>=3.1.0
 imutils~=0.5.4
-matplotlib~=3.1.3
+matplotlib
 mediapipe~=0.8.6
 opencv-contrib-python~=3.4.14.51 
 pandas~=1.2.5
+pyqtgraph~=0.12.1
 rich~=10.4.0
+scipy~=1.7.0
+toml~=0.10.2
 tqdm~=4.61.1
-ruamel.yaml<0.17
+tensorflow-gpu>=2.0
+pyqt5>=5.15.4
+pyopengl>=3.1.5
+tf_slim>=1.1.0
+deeplabcut>=2.2
+ruamel.yaml>=0.15.0
+typed-ast<1.5,>=1.4.0
 numpy<=1.19
+ffmpeg
