+++ setup.cfg
@@ -1,0 +1,6 @@
+[versioneer]
+VCS = git
+style = pep440-pre
+versionfile_source = freemocap/_version.py
+versionfile_build = freemocap/_version.py
+tag_prefix = v
