+++ docs/source/index.rst
@@ -1,0 +1,15 @@
+.. Packaging Scientific Python documentation master file, created by
+   sphinx-quickstart on Thu Jun 28 12:35:56 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+The FreeMoCap Project Documentation
+===================================
+
+.. toctree::
+   :maxdepth: 2
+
+   installation
+   usage
+   release-history
+   min_versions
