+++ AUTHORS.rst
@@ -1,0 +1,15 @@
+=======
+Credits
+=======
+
+Maintainers
+----------
+
+* Jonathan Samir Matthis <jonmatthis@gmail.com>
+Aaron Charian @aaroncherian
+
+Contributors
+------------
+
+
+None yet. Why not be the first? See: CONTRIBUTING.rst
