+++ MANIFEST.in
@@ -13,7 +13,6 @@ include versioneer.py
 include freemocap/_version.py
 
 include freemocap/logo/*
-
 # If including data files in the package, add them like:
 # include path/to/data_file
 
