+++ freemocap_runme_script.py
@@ -6,13 +6,13 @@ import sys
 fmc.RunMe(
     # sessionID = 'sesh_2021-08-29_17_55_45',
     sessionID=None,
-    stage = 5,
-    useOpenPose=True, 
+    stage = 1,
+    useOpenPose=False, 
     runOpenPose=True, 
-    useMediaPipe=False,
+    useMediaPipe=True,
     runMediaPipe = True,
     useDLC=False, 
     recordVid = True ,
-    reconstructionConfidenceThreshold = 0.0,
-    calVideoFrameLength = 60,
+    reconstructionConfidenceThreshold = 0.7,
+    calVideoFrameLength = -1,
 )
