### 2021-09-13 aa498f0
set runMediaPipe default to True and ignore user_prefs
