+++ freemocap/fmc_runme.py
@@ -27,8 +27,8 @@ def RunMe(sessionID=None,
         useOpenPose=False, 
         runOpenPose = True, 
         useMediaPipe=True,
+        runMediaPipe=True,
         useDLC=False,
-        runMediaPipe=False,
         dlcConfigPath=None,
         debug=False,
         setDataPath = False,
