+++ freemocap/user_preferences.yaml
@@ -1,18 +1,18 @@
 default:
-  path_to_save: C:\Users\jonma\Dropbox\GitKrakenRepos\freemocap
+  path_to_save: C:\
   dlc_config_paths: []
   rotations: {}
   parameters:
-    exposure: -5
+    exposure: -6
     resWidth: 640
     resHeight: 480
-    framerate: 25
+    framerate: 30
     codec: DIVX
 saved:
-  path_to_save: C:/Users/jonma/Dropbox/FreeMoCapProject
+  path_to_save: C:/
   rotations:
     0: 0
-    2: 0
+    1: 0
   parameters:
     exposure: -7
     resWidth: 640
