### 2021-09-13 9b6a00d
add switch to use CAP_DSHOW on windwos and CAP_ANY otherwise
Ostensibly for ubuntu (mac?) compatibility, see https://github.com/jonmatthis/freemocap/issues/52
