+++ freemocap/user_preferences.yaml
@@ -9,10 +9,10 @@ default:
     framerate: 30
     codec: DIVX
 saved:
-  path_to_save: C:/
+  path_to_save: C:/Users/jonma/Dropbox/FreeMoCapProject
   rotations:
     0: 0
-    1: 0
+    2: 0
   parameters:
     exposure: -7
     resWidth: 640
