### 2021-09-13 9440ac0
add switch to use CAP_DSHOW on windwos and CAP_ANY otherwise
(same as last commit, but I'd forgotten to save a few files lol)
