+++ update_pip.txt
@@ -1,8 +1,8 @@
 
-git commit --allow-empty -m "REL: v0.0.29"
-git tag -a v0.0.29  # Don't forget the leading v
+git commit --allow-empty -m "REL: v0.0.31"
+git tag -a v0.0.31  # Don't forget the leading v
 git push origin main
-git push origin v0.0.29
+git push origin v0.0.31
 python3 -m pip install --upgrade wheel
 git clean -dfx
 python setup.py sdist
