### 2021-09-13 931f864
tweaks to make mediapipe processing more accurate and plotting more robust
at the cost of speed, but in this house we tend to like our Speed/Accuracy dial turned all the way towards accuracy \o/
