### 2021-09-13 f6f7ec4
switching mediapipe's 'static_image_mode' back to false
I think it's better? 

Still has the problem of getting stuck on 'ghost skeleton' solutions, but i think this leads to better performance when a good skelly IS detected
