+++ freemocap/fmc_startup/user_preferences.yaml
@@ -1,21 +1,0 @@
-default:
-  path_to_save: C:\Users\aaron\Documents\GitHub\freemocap
-  dlc_config_paths: []
-  rotations: {}
-  parameters:
-    exposure: -6
-    resWidth: 640
-    resHeight: 480
-    framerate: 30
-    codec: DIVX
-saved:
-  path_to_save: C:\Users\aaron\Documents\GitHub\freemocap
-  rotations:
-    0: 0
-    1: 0
-  parameters:
-    exposure: -5
-    resWidth: 640
-    resHeight: 480
-    framerate: 25
-    codec: DIVX
