+++ freemocap/webcamGUI.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from ruamel.yaml import YAML
 import datetime
 import sys
 import pandas as pd
