+++ runme_FreeMoCap.py
@@ -3,12 +3,5 @@ import freemocap as fmc
 
 #fmc.RunMe(useOpenPose = False, useMediaPipe=True, stage = 5, sessionID ='sesh_2021-09-13_14_34_12', recordVid=False)
 
-fmc.RunDemo()
+fmc.RunMe()
 
-#fmc.RunMe()
-
-# word = 'data'
-
-# mystring = 'freemocap_data'
-# if word in mystring: 
-#    print('success')
