+++ requirements.txt
@@ -11,3 +11,4 @@ ruamel.yaml>=0.15.0
 numpy<=1.19
 typed-ast<1.5,>=1.4.0
 moviepy
+ipython
