### 2021-09-30 6217f36
change default to use the whole rec for calibration
