+++ update_pip.txt
@@ -8,3 +8,4 @@ git clean -dfx
 python setup.py sdist
 python setup.py bdist_wheel
 twine upload dist/*
+
