+++ requirements.txt
@@ -1,4 +1,5 @@
 aniposelib~=0.4.1
+numba~=0.50.1
 h5py>=3.1.0
 imutils~=0.5.4
 matplotlib~=3.1.3
