+++ freemocap-0.0.33/freemocap/fmc_mediapipe.py
@@ -1,372 +1,0 @@
-from pathlib import Path
-
-
-import numpy as np
-from rich.progress import Progress
-from ruamel.yaml import YAML
-import cv2
-import mediapipe as mp
-
-# from numba import jit
-
-mp_drawing = mp.solutions.drawing_utils
-mp_holistic = mp.solutions.holistic
-
-
-def runMediaPipe(session, dummyRun=False):
-    """ 
-    Run MediaPipe on synced videos, and save body tracking data to be parsed 
-    """  
-
-    session.mediaPipeDataPath.mkdir(exist_ok = True)
-
-    mediaPipe_jsonPathList = []  # list to hold the paths to the json files
-    mediaPipe_imgPathList = []
-    mediaPipe_imgPathList_yaml = []
-    mediaPipe_jsonPathList_yaml = []
-
-    with mp_holistic.Holistic() as holistic:
-        eachCamerasData = []  # Create an empty list that holds each cameras data
-        for (
-            thisVidPath
-        ) in (
-            session.syncedVidPath.iterdir()
-        ):  # Run MediaPipe 'Holistic' (body, hands, face) tracker on each video in the raw video folder
-            if (
-                thisVidPath.suffix == ".mp4"
-            ):  # NOTE - at some point we should build some list of 'synced video names' and check against that
-
-                vidPath = session.mediaPipeDataPath / thisVidPath.stem
-                jsonPath = vidPath / "json"
-                jsonPath.mkdir(
-                    parents=True, exist_ok=True
-                )  # this camera's json files (with keypoints)
-                imgPath = vidPath / "images"
-                imgPath.mkdir(parents=True, exist_ok=True)
-                mediaPipe_jsonPathList.append(jsonPath)
-                mediaPipe_imgPathList.append(imgPath)
-                mediaPipe_imgPathList_yaml.append(str(imgPath))
-                mediaPipe_jsonPathList_yaml.append(str(jsonPath))
-                mediaPipe_dataList = []  # Create an empty list for mediapipes data
-                if not dummyRun:
-                    cap = cv2.VideoCapture(str(thisVidPath))
-
-                    frameNum = -1
-                    numFrames = cap.get(cv2.CAP_PROP_FRAME_COUNT)
-                    success, image = cap.read()  # load first image from video
-
-                    with Progress() as progress:
-                        progressBar = progress.add_task(
-                            "[magenta]MediaPiping: {}".format(thisVidPath.name),
-                            total=numFrames,
-                        )  # make progressbar
-
-                        while success:
-
-                            if frameNum % 5 == 0:
-                                progress.update(
-                                    progressBar, advance=5
-                                )  # increment progress bar everyh 5th frame
-
-                            frameNum += 1
-
-                            image_height, image_width, _ = image.shape
-
-                            results = holistic.process(
-                                cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-                            )  # THIS IS WHERE THE MAGIC HAPENS (ps. Convert the BGR image to RGB before processing.)
-
-                            try:
-                                results.pose_landmarks.landmark
-                                try:
-                                    results.left_hand_landmarks.landmark
-                                    try:
-                                        results.right_hand_landmarks.landmark
-                                        try:
-                                            results.face_landmarks.landmark
-                                            f = 9
[…]
