+++ requirements.txt
@@ -11,5 +11,4 @@ tqdm~=4.61.1
 ruamel.yaml>=0.15.0
 numpy<=1.19
 typed-ast<1.5,>=1.4.0
-moviepy
-ipython
+moviepy ~=1.0.3
