+++ freemocap/fmc_startup/user_preferences.yaml
@@ -1,11 +1,0 @@
-default:
-  path_to_save: C:\Users\jonma\Dropbox\GitKrakenRepos\freemocap
-  dlc_config_paths: []
-  rotations: {}
-  parameters:
-    exposure: -5
-    resWidth: 640
-    resHeight: 480
-    framerate: 25
-    codec: DIVX
-saved: {}
