+++ runme_FreeMoCap.py
@@ -3,5 +3,5 @@ import freemocap as fmc
 
 #fmc.RunMe(useOpenPose = False, useMediaPipe=True, stage = 5, sessionID ='sesh_2021-09-13_14_34_12', recordVid=False)
 
-fmc.RunMe()
+fmc.RunMe(sessionID = 'sesh_2021-08-29_17_55_45',stage = 3)
 
