### 2021-10-15 eea1115
fix forever long rich traceback
when show locals is set to True, it prints out the charuco points array and its huge
