+++ freemocap/fmc_mediapipe.py
@@ -9,6 +9,7 @@ import mediapipe as mp
 # from numba import jit
 
 mp_drawing = mp.solutions.drawing_utils
+mp_drawing_styles = mp.solutions.drawing_styles
 mp_holistic = mp.solutions.holistic
 
 
