+++ freemocap_runme_script.py
@@ -1,3 +1,3 @@
 import freemocap as fmc
-fmc.RunMe()
+fmc.RunMe(sessionID = 'sesh_2021-10-19_18_40_12', stage=5, useOpenPose=True, setDataPath=False)
 
