### 2021-10-23 0b4262c
Stop making unnecessary MediaPipe and OpenPose stuff and stopped saving open pose images
