+++ freemocap/play_skeleton_animation.py
@@ -63,7 +63,7 @@ def PlaySkeletonAnimation(
     vidType=1,
     startFrame=0,
     azimuth=-90,
-    elevation=-61,
+    elevation=-70,
     useOpenPose=True,
     useMediaPipe=False,
     useDLC=False,
@@ -262,7 +262,9 @@ def PlaySkeletonAnimation(
             if not thisArtistKey == 'blackLine':
                 xCurrTimeArtists[thisArtistKey][0].set_ydata(xCurrTimeYdata[thisArtistKey][frameNum] ) 
                 yCurrTimeArtists[thisArtistKey][0].set_ydata(yCurrTimeYdata[thisArtistKey][frameNum] ) 
-                
+
+        yCurrTimeArtists['frameNumber'].set_x(frameNum/fps+3/fps) 
+        yCurrTimeArtists['frameNumber'].set_text('Frame# '+ str(frameNum))
         # if recordVid:
         #     thisFramePath = str(animationFramePath) + '/'+ str(session.sessionID) + "_frame_" + str(frameNum).zfill(6) + ".png"
         #     plt.savefig(thisFramePath)
@@ -532,7 +534,7 @@ def PlaySkeletonAnimation(
     # groundYY = np.zeros_like(groundXX)
     # groundMesh = ax3d.plot_surface(groundXX, groundZZ, groundYY, color='k', alpha=.5)
 
-    axRange = 750#session.board.square_length * 10
+    axRange = 400#session.board.square_length * 10
 
     # Setting the axes properties
     ax3d.set_xlim3d([mx-axRange, mx+axRange])
@@ -739,8 +741,9 @@ def PlaySkeletonAnimation(
     yCurrTimeYdata['rHandDot'] = rHandY
     yCurrTimeArtists['lHandDot']  = yTimeSeriesAx.plot([startFrame/fps], [lHandY[startFrame]], markeredgecolor=humon_blue, markerfacecolor = 'k', marker='o', markersize=3)
     yCurrTimeYdata['lHandDot'] = lHandY
-
-
+    
+    
+    yCurrTimeArtists['frameNumber'] = yTimeSeriesAx.text(startFrame/fps+3/fps, ylimRange*.7, "Frame# " + str(startFrame), fontsize=6)
 
     yTimeSeriesAx.tick_params(labelsize=6, direction='in', width=.5)
     yTimeSeriesAx.tick_params( pad=2)
@@ -807,7 +810,7 @@ def PlaySkeletonAnimation(
     # )
 
     # thisVidAxis.text(-565,1440, 'Music - artist: Neon Exdeath, song: Meowmaline, album: Jewel Tones',color=humon_green, fontsize=6)
-    thisVidAxis.text(-150,510, 'github.com/jonmatthis/freemocap || jonmatthis.com/freemocap',color=humon_green, fontsize=6)
+    thisVidAxis.text(0,thisVidCap.get(cv2.CAP_PROP_FRAME_HEIGHT)*1.06, 'github.com/jonmatthis/freemocap || freemocap.org',color=humon_green, fontsize=6)
 
     logoAx = fig.add_subplot(position=[.85, .85, .15, .15])
     logoIm = cv2.imread(r'logo\fmc-logo-black-border-white-bkgd.png') #JSM NOTE - THis doesn't work, but SOMETIMES IT DOES?!
