+++ runme_FreeMoCap.py
@@ -1,7 +1,4 @@
-from freemocap import session
 import freemocap as fmc
 
-#fmc.RunMe(useOpenPose = False, useMediaPipe=True, stage = 5, sessionID ='sesh_2021-09-13_14_34_12', recordVid=False)
-
-fmc.RunMe(useBlender=True)
+fmc.RunMe()
 
