### 2021-10-23 ea8249a
Stop making unnecessary MediaPipe and OpenPose stuff and stopped saving open pose images
