### 2021-10-25 8ab62de
Added `FMC_MultiCameraRecorder` and `FMC_Camera` classes
Run each .py file for a demo
