+++ freemocap/dev_scratchpad/dev_alpha/scratch/VideoTemplate_pyqt5.py
@@ -1,0 +1,209 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '.\VideoTemplate.ui'
+#
+# Created by: PyQt5 UI code generator 5.15.4
+#
+# WARNING: Any manual changes made to this file will be lost when pyuic5 is
+# run again.  Do not edit this file unless you know what you are doing.
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class Ui_MainWindow(object):
+    def setupUi(self, MainWindow):
+        MainWindow.setObjectName("MainWindow")
+        MainWindow.resize(695, 798)
+        self.centralwidget = QtWidgets.QWidget(MainWindow)
+        self.centralwidget.setObjectName("centralwidget")
+        self.gridLayout_2 = QtWidgets.QGridLayout(self.centralwidget)
+        self.gridLayout_2.setObjectName("gridLayout_2")
+        self.cudaCheck = QtWidgets.QCheckBox(self.centralwidget)
+        self.cudaCheck.setObjectName("cudaCheck")
+        self.gridLayout_2.addWidget(self.cudaCheck, 9, 0, 1, 2)
+        self.numbaCheck = QtWidgets.QCheckBox(self.centralwidget)
+        self.numbaCheck.setObjectName("numbaCheck")
+        self.gridLayout_2.addWidget(self.numbaCheck, 9, 2, 1, 2)
+        self.downsampleCheck = QtWidgets.QCheckBox(self.centralwidget)
+        self.downsampleCheck.setObjectName("downsampleCheck")
+        self.gridLayout_2.addWidget(self.downsampleCheck, 8, 0, 1, 2)
+        self.scaleCheck = QtWidgets.QCheckBox(self.centralwidget)
+        self.scaleCheck.setObjectName("scaleCheck")
+        self.gridLayout_2.addWidget(self.scaleCheck, 4, 0, 1, 1)
+        self.gridLayout = QtWidgets.QGridLayout()
+        self.gridLayout.setObjectName("gridLayout")
+        self.rawRadio = QtWidgets.QRadioButton(self.centralwidget)
+        self.rawRadio.setObjectName("rawRadio")
+        self.gridLayout.addWidget(self.rawRadio, 3, 0, 1, 1)
+        self.gfxRadio = QtWidgets.QRadioButton(self.centralwidget)
+        self.gfxRadio.setChecked(True)
+        self.gfxRadio.setObjectName("gfxRadio")
+        self.gridLayout.addWidget(self.gfxRadio, 2, 0, 1, 1)
+        self.stack = QtWidgets.QStackedWidget(self.centralwidget)
+        self.stack.setObjectName("stack")
+        self.page = QtWidgets.QWidget()
+        self.page.setObjectName("page")
+        self.gridLayout_3 = QtWidgets.QGridLayout(self.page)
+        self.gridLayout_3.setObjectName("gridLayout_3")
+        self.graphicsView = GraphicsView(self.page)
+        self.graphicsView.setObjectName("graphicsView")
+        self.gridLayout_3.addWidget(self.graphicsView, 0, 0, 1, 1)
+        self.stack.addWidget(self.page)
+        self.page_2 = QtWidgets.QWidget()
+        self.page_2.setObjectName("page_2")
+        self.gridLayout_4 = QtWidgets.QGridLayout(self.page_2)
+        self.gridLayout_4.setObjectName("gridLayout_4")
+        self.rawImg = RawImageWidget(self.page_2)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Preferred)
[…]
