### 2021-10-25 8b85da2
Added `FMC_MultiCameraRecorder` and `FMC_Camera` classes
Run each .py file for a demo
