+++ freemocap/dev_scratchpad/dev_alpha/fmc_camera.py
@@ -1,0 +1,333 @@
+##
+##
+##  ███████ ███    ███  ██████          ██████  █████  ███    ███ ███████ ██████   █████  
+##  ██      ████  ████ ██              ██      ██   ██ ████  ████ ██      ██   ██ ██   ██ 
+##  █████   ██ ████ ██ ██              ██      ███████ ██ ████ ██ █████   ██████  ███████ 
+##  ██      ██  ██  ██ ██              ██      ██   ██ ██  ██  ██ ██      ██   ██ ██   ██ 
+##  ██      ██      ██  ██████          ██████ ██   ██ ██      ██ ███████ ██   ██ ██   ██ 
+##
+##
+#Font - ANSI Regular - https://patorjk.com/software/taag/#p=display&f=ANSI%20Regular&t=Play%20Skeleton%20Animation
+                                                                                      
+                                                                                      
+
+import platform
+from pathlib import Path
+import time
+
+import cv2
+import numpy as np
+
+#python rich stuff - https://rich.readthedocs.io/
+from rich import print
+from rich.console import Console
+from rich.live import Live
+from rich.table import Table    
+
+
+cap_default_parameters_dict = {
+                            'exposure': -6,
+                            'cap_resolution_width': 640,
+                            'cap_resolution_height': 480,
+                            }
+
+class FMC_Camera:
+    """Simple class to open a cv2.VideoCapture object, with parts to connect to a `MutliCamRecorder` object
+    """
+    def __init__(
+                self, 
+                cam_num=0, 
+                cap_parameters_dict = cap_default_parameters_dict,
+                vid_save_path=None,  
+                frame_queue = None,
+                thread_barrier = None,
+                exit_event = None,
+                show_cam_stream = False,  
+                console = None,       
+                show_console = False,       
+                ):
+        """Open a camera stream using `cv2.VideoCapture()`. 
+        If `Windows`, use `cv2.VideoCapture(camNum, cv2.CAP_DSHOW)` to make video initialize much faster. Otherwise, use `cv2.VideoCapture(camNum, cv2.CAP_ANY)
+
+        Args:
+            camNum (int, optional): The ID of the camera to open. Defaults to 0 (i.e. the first camera it finds)
+            vid_save_path (pathlib Path, optional): Where to save the video. Defaults to None.
+            barrier (threading.Barrier, optional): Barrier for syncronized multicam recording. Defaults to None.
+            frame_queue (queue.Queue, optional): [description]. Queue to send out images when we're recording in a thread.
+        """
+
+        self._cam_num = cam_num
+        self._cam_name = 'Camera_{}'.format(str(self._cam_num).zfill(2))
+        self._cam_short_name = 'cam{}'.format(self._cam_name)
+        
+        self._vid_cap_start_time_unix = None
+
+        self._cap_parameters_dict = cap_parameters_dict
+
+        self._frame_queue = frame_queue
+        self._thread_barrier = thread_barrier
+        self._exit_event = exit_event
+        self._vid_save_path = vid_save_path
+        self._show_console = show_console
+        self.rich_console = console # a console object from the Rich python package
+
+
+        if vid_save_path:
+            self._vid_save_path = Path(vid_save_path)
+        
+
+        if not console:
+            self.rich_console = Console()
+        
+        self.open() #open VideoCapture Object
+
+        self._show_cam_stream = show_cam_stream
+        if self._show_cam_stream:
+            self.show()#display camera stream
+
+        if self._frame_queue: 
+            self.run_in_thread() #launch in thread (or process?) mode
+    ##   
[…]
