+++ freemocap/dev_scratchpad/dev_alpha/FMC_MultiCameraRecorder.py
@@ -1,0 +1,287 @@
+from os import mkdir
+from fmc_camera import FMC_Camera
+
+import concurrent.futures
+import logging
+import queue
+import threading
+import time
+import platform
+import datetime
+
+import cv2
+import matplotlib.pyplot as plt
+import numpy as np
+import h5py
+from pathlib import Path
+
+from rich import print
+from rich import inspect
+from rich.progress import track
+from rich.traceback import install
+install(show_locals=True)
+from rich.console import Console
+from rich.progress import track
+
+
+class FMC_MultiCameraRecorder:
+    
+    def __init__(
+                self,
+                rec_name = None,
+                save_path = None,
+                num_cams = 0,
+                cams_to_use_list = None,
+                show_multi_cam_stream = True,
+                save_to_mp4 = True,
+                save_to_h5 = False,
+                save_log_file = False
+                ):
+        
+        self._init_start_time = time.time_ns()#the precision is aspirational, lol
+
+        if rec_name is None:
+            self._rec_name = datetime.datetime.now().strftime("FMC_MultiCamRecording_%Y-%b-%d_%H_%M_%S")
+        else:
+            self._rec_name = rec_name
+
+
+        if save_path is None:
+            self._save_path = Path('data/' + self._rec_name)
+        else:
+            self._save_path = Path(save_path)
+        
+        if save_path or save_to_mp4 or save_to_h5 or save_log_file:
+            self._save_path.mkdir(parents=True)
+            self._path_to_log_file = Path(str(self._save_path / self._rec_name) + '_log.txt')
+
+        self._save_to_mp4 = save_to_mp4
+        if self._save_to_mp4:            
+            self._output_video_object = None #build this after we get the first multi_frame_image
+
+        # log_file_obj = open(self._path_to_log_file,'w+')
+        # self.rich_console = Console(file=log_file_ob)
+        self.rich_console = Console()
+
+        self._num_cams = num_cams
+        self._cams_to_use_list = cams_to_use_list
+        
+        self.exit_event = threading.Event()
+
+        
+        
+        self.rich_console.rule('Starting FreeMoCap MultiCam Recorder!')
+
+        if self.cams_to_use_list is None:
+            self.find_available_cameras()
+        
+
+        self.start_multi_cam_thread_pool()
+    
+    @property
+    def num_cams(self):
+        """How many cameras to use"""        
+        return self._num_cams
+
+    @num_cams.setter
+    def num_cams(self, num_cams_input):
+        """Define how many cameras to use"""
+        self._num_cams = num_cams_input        
+
+    @property
+    def cams_to_use_list(self):
+        """ a list denoting which camera IDs to use.
+         If both this and `num_cams` are set to default value (`0` and `None`, respectively), use all available cameras).
+         """
+        return self._cams_to_use_list
+
+    @property
+    def init_start_time(self):
+        """Time that this object was initialized (UNIX Epoch time from `time.time_ns())
+
+        Returns:
+            float: time this object was created (UNIX Epoch time from `time.time_ns())
+        """
+        return self._init_start_time
+    
+    @property
+    def each_cam_timestamps_unix_ns(self): #not really a property though?
+        return self._each_cam_timestamps_unix_ns
+
+    
+    def __enter__(self):
+        """Context manager -  No need to do anything special on start"""
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
[…]
