### 2021-10-25 bb077c0
added block text to ease navigation and scrolling
