+++ freemocap/dev_scratchpad/dev_alpha/FMC_MultiCameraRecorder.py
@@ -1,3 +1,14 @@
+##   
+##   
+##   ███████ ███    ███  ██████     ███    ███ ██    ██ ██      ████████ ██        ██████  █████  ███    ███ ███████ ██████   █████  
+##   ██      ████  ████ ██          ████  ████ ██    ██ ██         ██    ██       ██      ██   ██ ████  ████ ██      ██   ██ ██   ██ 
+##   █████   ██ ████ ██ ██          ██ ████ ██ ██    ██ ██         ██    ██ █████ ██      ███████ ██ ████ ██ █████   ██████  ███████ 
+##   ██      ██  ██  ██ ██          ██  ██  ██ ██    ██ ██         ██    ██       ██      ██   ██ ██  ██  ██ ██      ██   ██ ██   ██ 
+##   ██      ██      ██  ██████     ██      ██  ██████  ███████    ██    ██        ██████ ██   ██ ██      ██ ███████ ██   ██ ██   ██ 
+##                                                                                                                                   
+##                                                                                                                                   
+
+
 from os import mkdir
 from fmc_camera import FMC_Camera
 
@@ -25,7 +36,16 @@ from rich.progress import track
 
 
 class FMC_MultiCameraRecorder:
-    
+    ##  
+    ##  
+    ##                  ██ ███    ██ ██ ████████                 
+    ##                  ██ ████   ██ ██    ██                    
+    ##                  ██ ██ ██  ██ ██    ██                    
+    ##                  ██ ██  ██ ██ ██    ██                    
+    ##  ███████ ███████ ██ ██   ████ ██    ██    ███████ ███████ 
+    ##                                                           
+    ##  
+    ##  
     def __init__(
                 self,
                 rec_name = None,
@@ -78,6 +98,16 @@ class FMC_MultiCameraRecorder:
 
         self.start_multi_cam_thread_pool()
     
+    ###   
+    ###   
+    ###   ██████  ██████   ██████  ██████  ███████ ██████  ████████ ██ ███████ ███████ 
+    ###   ██   ██ ██   ██ ██    ██ ██   ██ ██      ██   ██    ██    ██ ██      ██      
+    ###   ██████  ██████  ██    ██ ██████  █████   ██████     ██    ██ █████   ███████ 
+    ###   ██      ██   ██ ██    ██ ██      ██      ██   ██    ██    ██ ██           ██ 
+    ###   ██      ██   ██  ██████  ██      ███████ ██   ██    ██    ██ ███████ ███████ 
+    ###                                                                                
+    ###                                                                                
+
     @property
     def num_cams(self):
         """How many cameras to use"""        
@@ -107,16 +137,16 @@ class FMC_MultiCameraRecorder:
     @property
     def each_cam_timestamps_unix_ns(self): #not really a property though?
         return self._each_cam_timestamps_unix_ns
+    ###   
+    ###   
+    ###   ███    ███ ███████ ████████ ██   ██  ██████  ██████  ███████ 
+    ###   ████  ████ ██         ██    ██   ██ ██    ██ ██   ██ ██      
+    ###   ██ ████ ██ █████      ██    ███████ ██    ██ ██   ██ ███████ 
+    ###   ██  ██  ██ ██         ██    ██   ██ ██    ██ ██   ██      ██ 
+    ###   ██      ██ ███████    ██    ██   ██  ██████  ██████  ███████ 
+    ###                                                                
+    ###                                                                    
 
-    
-    def __enter__(self):
-        """Context manager -  No need to do anything special on start"""
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        """Context manager -  on exit, set `self.exit_event` to activate shutdown sequence """
-        self.create_diagnostic_images()
-        return self
 
[…]
