+++ freemocap/dev_scratchpad/dev_alpha/FMC_MultiCameraRecorder.py
@@ -10,12 +10,14 @@
 
 
 from os import mkdir
-from fmc_camera import FMC_Camera
+from FMC_Camera import FMC_Camera
 
 import concurrent.futures
 import logging
 import queue
 import threading
+import pathos.multiprocessing as pathos_mp
+from pathos.helpers import mp as pathos_mp_helper
 import time
 import platform
 import datetime
@@ -32,6 +34,7 @@ from rich.progress import track
 from rich.traceback import install
 install(show_locals=True)
 from rich.console import Console
+rich_console = Console()
 from rich.progress import track
 
 
@@ -80,24 +83,20 @@ class FMC_MultiCameraRecorder:
             self._output_video_object = None #build this after we get the first multi_frame_image
 
         # log_file_obj = open(self._path_to_log_file,'w+')
-        # self.rich_console = Console(file=log_file_ob)
-        self.rich_console = Console()
+        # rich_console = Console(file=log_file_obj)
+        rich_console = Console()
 
         self._num_cams = num_cams
         self._cams_to_use_list = cams_to_use_list
         
-        self.exit_event = threading.Event()
-
         
         
-        self.rich_console.rule('Starting FreeMoCap MultiCam Recorder!')
+        
+        rich_console.rule('Starting FreeMoCap MultiCam Recorder!')
 
-        if self.cams_to_use_list is None:
-            self.find_available_cameras()
+
         
 
-        self.start_multi_cam_thread_pool()
-    
     ###   
     ###   
     ###   ██████  ██████   ██████  ██████  ███████ ██████  ████████ ██ ███████ ███████ 
@@ -146,7 +145,18 @@ class FMC_MultiCameraRecorder:
     ###   ██      ██ ███████    ██    ██   ██  ██████  ██████  ███████ 
     ###                                                                
     ###                                                                    
+    def start(self):
 
+        if self.cams_to_use_list is None:
+            self.find_available_cameras()
+        
+        # pathos_pool = ProcessPool(nodes=1)
+        # pathos_pool.map(self.start_multi_cam_thread_pool)
+        # self.start_multi_cam_thread_pool() #threads are mostly fine, but I think processes will be better eventually? 
+        self.start_multi_cam_process_pool() # as of 2021-10-26, this doesn't work :( 
+        self.create_diagnostic_images()
+
+    
 
     def find_available_cameras(self):
         """find available webcams and return IDs in list."""
@@ -167,7 +177,7 @@ class FMC_MultiCameraRecorder:
             if cap.isOpened():
                 self._cams_to_use_list.append(camNum)
             cap.release()
-        console.print("Found cameras at ", self._cams_to_use_list)
+        rich_console.print("Found cameras at ", self._cams_to_use_list)
         
         self.num_cams = len(self._cams_to_use_list)
         return self.cams_to_use_list
@@ -189,18 +199,65 @@ class FMC_MultiCameraRecorder:
         self.cam_frame_queue = queue.Queue(maxsize=self.num_cams)
         self.multi_cam_image_queue = queue.Queue()
 
+        self.exit_event = threading.Event()
+
         self._each_cam_timestamps_unix_ns = None
         self._cam_thread = [None] * self.num_cams
 
-        self.rich_console.rule('starting multi_cam_thread_pool')
+        rich_console.rule('starting multi_cam_thread_pool')
 
         while not self.exit_event.is_set():
             with concurrent.futures.ThreadPoolExecutor(max_workers=self.num_cams+1) as executor:
[…]
