### 2021-10-28 dfffc40
added print statements when saving out files
