+++ freemocap_runme_script.py
@@ -1,3 +1,3 @@
 import freemocap as fmc
-fmc.RunMe(sessionID = 'sesh_2021-10-15_18_09_48_goPro', stage=5,  )
+fmc.RunMe( )
 
