### 2021-10-28 090f833
mediapipe overlay will only show up in setup window
