+++ freemocap/fmc_runme.py
@@ -49,6 +49,7 @@ def RunMe(sessionID=None,
         setDataPath = False,
         userDataPath = None,
         recordVid = True,
+        showAnimation = True,
         reconstructionConfidenceThreshold = .7,
         charucoSquareSize = 36,#mm - ~the size of the squares when printed on 8.5x11" paper based on parameters in ReadMe.md
         calVideoFrameLength = -1,
@@ -109,6 +110,10 @@ def RunMe(sessionID=None,
                         marker_bits=4, dict_size=250)
     sesh.board = board
 
+    console.rule()    
+    console.rule('Finding available webcams')
+    console.rule()    
+
     # %% Initialization
     if stage == 1:
         camera_settings.initialize(sesh,stage,board)
@@ -119,18 +124,22 @@ def RunMe(sessionID=None,
 
     # %% Stage One
     if stage <= 1:
-        print()
-        print('Starting Video Recordings')
+        console.rule()    
+        console.rule('Starting Video Recordings')
+        console.rule()    
+
         runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
         sesh.save_session()
     else:
-        print('Skipping Video Recording')
+         console.rule('Skipping Video Recording')
 
 
     # %% Stage Two
     if stage <= 2:
-        print()
-        print('Starting Video Syncing')
+        console.rule()    
+        console.rule('Synchronizing Recorded Videos')
+        console.rule()    
+
         runcams.SyncCams(sesh, sesh.timeStampData,sesh.numCamRange,sesh.vidNames,sesh.camIDs)
         sesh.save_session()
     else:
@@ -138,17 +147,30 @@ def RunMe(sessionID=None,
 
     # %% Stage Three
     if stage <= 3:
-        print()
-        print('Starting Calibration')
-        sesh.cgroup, sesh.mean_charuco_fr_mar_xyz = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength)
+        console.rule()    
+        console.rule('Starting Anipose Calibration')
+        console.rule('Identifying 6 DoF position of each camera based on detected charuco boards')
+        console.rule()    
+
+        print('See https://anipose.org for details')
+        sesh.cgroup, sesh.mean_charuco_fr_mar_xyz = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength, spoof_anipose_fail_bool=False)
+        print('Anipose Calibration Successful!')
     else:
         print('Skipping Calibration')
 
     # %% Stage Four
     if stage <= 4:
+        console.rule()    
+        console.rule('Starting 2D Point Trackers')
+        console.rule('This step impolements the machine learning driven computer vision convolutional neural networks that track the skeleton, etc')
+        console.rule('This part is crazy future tech sci fi stuff. Seriously unbelievable this kind of thing is possible.')
+        console.rule()    
 
-        print('Starting Track Image Points')
         if sesh.useMediaPipe:
+            console.rule()
+            console.rule('Running MediaPipe skeleton tracker - https://google.github.io/mediapipe')
+            console.rule()
+
             if runMediaPipe:
                 fmc_mediapipe.runMediaPipe(sesh)
 
@@ -159,6 +181,10 @@ def RunMe(sessionID=None,
         sesh.save_session()
 
         if sesh.useOpenPose:
+            console.rule()
+            console.rule('Running OpenPose skeleton tracker - https://github.com/CMU-Perceptual-Computing-Lab/openpose')
+            console.rule()
+
             fmc_openpose.runOpenPose(sesh, runOpenPose=runOpenPose)
[…]
