### 2021-10-30 0d40d6b
Added prettier Rich prints - And iterative Anipose solver
Made the print statements in the `runme` file prettier. Next, I'd love to color code the 'rule' lines to denote new 'stages'

Also - The 'calibrate' code will now iteratively re-try to run Anipose's calibration method with different window sizes until it either gets a success or shrinks window_size to less than 1 frame
