### 2021-10-30 097b451
setting cv2.VideoCapture to use 'cv2.CAP_DSHOW' for windows and 'cv2.CAP_ANY' for others
in an attempt at Ubuntu and Mac support
