### 2021-10-30 e5f265b
setting cv2.VideoCapture to use 'cv2.CAP_DSHOW' for windows and 'cv2.CAP_ANY' for others
in an attempt at Ubuntu and Mac support
