### 2021-10-31 f1284fe
prettying up rich output ✨
