+++ freemocap/fmc_runme.py
@@ -5,7 +5,7 @@ from freemocap.webcam import camera_settings, timesync
 from pathlib import Path
 import os
 import subprocess
-
+import time 
 from aniposelib.boards import CharucoBoard
 
 import numpy as np
@@ -20,6 +20,7 @@ from rich.markdown import Markdown
 from rich.traceback import install
 install(show_locals=False)
 from rich import inspect
+from rich.padding import Padding
 
 
 from freemocap import (
@@ -36,6 +37,7 @@ from freemocap import (
 
 
 
+thisStage = 0 #global
 
 def RunMe(sessionID=None,
         stage=1,
@@ -55,7 +57,8 @@ def RunMe(sessionID=None,
         calVideoFrameLength = -1,
         startFrame = 0,
         useBlender = False,
-        resetBlenderExe = False
+        resetBlenderExe = False,
+        pauseBetweenStages =1
         ):
     """ 
     Starts the freemocap pipeline based on either user-input values, or default values. Creates a new session class instance (called sesh)
@@ -92,8 +95,9 @@ def RunMe(sessionID=None,
             subfolders = [f.path for f in os.scandir(sesh.dataFolderPath) if f.is_dir()]  # copy-pasta from who knows where
             sesh.sessionID = Path(subfolders[-1]).stem  # grab the name of the last folder in the list of subfolders
         
+        console.rule()
         print('Running ' + str(sesh.sessionID) + ' from ' + str(sesh.dataFolderPath))
-
+        console.rule()
     if useBlender == True:
         here = Path(__file__).parent
         subprocessPath = here/'fmc_blender.py'
@@ -110,9 +114,7 @@ def RunMe(sessionID=None,
                         marker_bits=4, dict_size=250)
     sesh.board = board
 
-    console.rule()    
-    console.rule('Finding available webcams')
-    console.rule()    
+   
 
     # %% Initialization
     if stage == 1:
@@ -123,23 +125,26 @@ def RunMe(sessionID=None,
         sesh.initialize(stage)
 
     # %% Stage One
+    
     if stage <= 1:
-        console.rule()    
-        console.rule('Starting Video Recordings')
-        console.rule()    
+        thisStage=1
+        console.rule(style="color({})".format(thisStage))    
+        console.rule('Starting Video Recordings'.upper(), style="color({})".format(thisStage))   
+        console.rule(style="color({})".format(thisStage))    
 
         runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
         sesh.save_session()
     else:
-         console.rule('Skipping Video Recording')
+        print('Skipping Video Recording')
 
 
     # %% Stage Two
     if stage <= 2:
-        console.rule()    
-        console.rule('Synchronizing Recorded Videos')
-        console.rule()    
-
+        thisStage=2
+        console.rule(style="color({})".format(thisStage))        
+        console.rule('Synchronizing Recorded Videos'.upper(),style="color({})".format(thisStage))    
+        console.rule(style="color({})".format(thisStage))    
+        time.sleep(pauseBetweenStages)  
         runcams.SyncCams(sesh, sesh.timeStampData,sesh.numCamRange,sesh.vidNames,sesh.camIDs)
         sesh.save_session()
     else:
@@ -147,12 +152,14 @@ def RunMe(sessionID=None,
 
     # %% Stage Three
     if stage <= 3:
-        console.rule()    
-        console.rule('Starting Anipose Calibration')
-        console.rule('Identifying 6 DoF position of each camera based on detected charuco boards')
-        console.rule()    
+        thisStage=3
+        console.rule(style="color({})".format(thisStage))    
[…]
