### 2021-11-02 be0269d
remove Anipose iterator - Doesn't quite work...
Something in Anipose spawns a new Thread which makes it able to jump out of the try/except block somehow?!
