+++ freemocap/calibrate.py
@@ -16,7 +16,7 @@ from rich.console import Console
 console = Console()
 
 
-def CalibrateCaptureVolume(session,board, calVideoFrameLength = .25, spoof_anipose_fail_bool=False):
+def CalibrateCaptureVolume(session,board, calVideoFrameLength = .25):
     """ 
     Check if a previous calibration yaml exists, and if not, create a set of shortened calibration videos and run Anipose functions
     to create a calibration yaml. Takes the 2D charuco board points and reconstructs them into 3D points that are saved out
@@ -29,17 +29,19 @@ def CalibrateCaptureVolume(session,board, calVideoFrameLength = .25, spoof_anipo
     session.dataArrayPath.mkdir(exist_ok = True)
     calibrationVideoPath = session.calVidPath
 
-    if calVideoFrameLength < 0: # if '-1' use the whole video
-        calVideoFrameLength = [0, session.numFrames]
-        calibrationVideoPath = session.calVidPath
-    else:
-        if calVideoFrameLength>0 and calVideoFrameLength<1: #if between 0 and 1, use as a percentage of the total video length
-            calVideoFrameLength = [0, round(calVideoFrameLength * session.numFrames)]
-        else: #otherwise, just use the input value as the number of frames to use        
-            calVideoFrameLength = [0, calVideoFrameLength]
-        createCalibrationVideos(session, calVideoFrameLength)
+    if type(calVideoFrameLength)==int or type(calVideoFrameLength)==float:
+        if calVideoFrameLength < 0: # if '-1' use the whole video
+            calVideoFrameLength = [0, session.numFrames]
+            calibrationVideoPath = session.calVidPath
+        else:
+            if calVideoFrameLength>0 and calVideoFrameLength<1: #if between 0 and 1, use as a percentage of the total video length
+                calVideoFrameLength = [0, round(calVideoFrameLength * session.numFrames)]
+            else: #otherwise, just use the input value as the number of frames to use        
+                calVideoFrameLength = [0, calVideoFrameLength]
 
-    calVideoFrameLength_og = copy.copy(calVideoFrameLength)
+    createCalibrationVideos(session, calVideoFrameLength)
+
+
 
     vidnames = []
     cam_names = []
@@ -54,49 +56,51 @@ def CalibrateCaptureVolume(session,board, calVideoFrameLength = .25, spoof_anipo
 
     session.cameraCalFilePath = session.sessionPath / calibrationFile
 
+    error,charuco_data, charuco_frames = cgroup.calibrate_videos(vidnames, board)
     
-    try:
-        error,charuco_data, charuco_frames = cgroup.calibrate_videos(vidnames, board)
-        if not spoof_anipose_fail_bool:
-            anipose_success = True        
-        else:
-            anipose_success = False
-    except:
-        anipose_success = False
-        console.rule('Anipose Calibration Failed for frame range {} - {}'.format(calVideoFrameLength[0], calVideoFrameLength[1]), style="color({})".format(3))
+    ## iteratively retry if anipose fails, butcouldn't get this to work because anipose launches some weird thread process that doesn't terminate properly? 
+    # try:
+    #     error,charuco_data, charuco_frames = cgroup.calibrate_videos(vidnames, board)
+    #     if not spoof_anipose_fail_bool:
+    #         anipose_success = True        
+    #     else:
+    #         anipose_success = False
+    # except:
+    #     anipose_success = False
+    #     console.rule('Anipose Calibration Failed for frame range {} - {}'.format(calVideoFrameLength[0], calVideoFrameLength[1]), style="color({})".format(3))
 
-    anipose_iter = -1
-    if not anipose_success: #%% run anipose calibration, and if it fails, re-make calibration vids with different frame range and try again
[…]
