### 2021-11-02 39c045f
Lots of aesthetic and 'usability' changes to print statements and labels
