+++ freemocap/fmc_runme.py
@@ -1,4 +1,5 @@
 
+from scipy.ndimage.measurements import center_of_mass
 from freemocap.fmc_startup import startup, startupGUI
 from freemocap.webcam import camera_settings, timesync
 
@@ -54,7 +55,7 @@ def RunMe(sessionID=None,
         showAnimation = True,
         reconstructionConfidenceThreshold = .7,
         charucoSquareSize = 36,#mm - ~the size of the squares when printed on 8.5x11" paper based on parameters in ReadMe.md
-        calVideoFrameLength = -1,
+        calVideoFrameLength = .5,
         startFrame = 0,
         useBlender = False,
         resetBlenderExe = False,
@@ -128,9 +129,9 @@ def RunMe(sessionID=None,
     
     if stage <= 1:
         thisStage=1
-        console.rule(style="color({})".format(thisStage))    
-        console.rule('Starting Video Recordings'.upper(), style="color({})".format(thisStage))   
-        console.rule(style="color({})".format(thisStage))    
+        console.rule(style="color({})".format(14))    
+        console.rule('Starting Video Recordings'.upper(), style="color({})".format(14))   
+        console.rule(style="color({})".format(14))    
 
         runcams.RecordCams(sesh, sesh.cam_inputs, sesh.parameterDictionary, sesh.rotationInputs)
         sesh.save_session()
@@ -156,11 +157,47 @@ def RunMe(sessionID=None,
         console.rule(style="color({})".format(thisStage))    
         console.rule('Starting Capture Volume Calibration'.upper(),style="color({})".format(thisStage))    
         console.rule(style="color({})".format(thisStage))  
-        console.print(Padding('Using Anipose to calculate 6 degree-of-freedom position (and distortion coeffs) of each camera based on detected charuco boards. This information is used to create a Camera Projection Matrix for each camera, which is later used in the 3d reconstruction stage', (1,4)), overflow="fold", justify='center')
+        console.print(Padding('Using Anipose to calculate 6 degree-of-freedom position (and distortion coeffs) of each camera based on detected charuco boards. This information is used to create a Camera Projection Matrix for each camera, which is later used in the 3d reconstruction stage', (1,4)), overflow="fold", justify='center',style="color({})".format(thisStage))
         console.rule('See https://anipose.org for details', style="color({})".format(thisStage))  
         console.rule(style="color({})".format(thisStage))    
 
-        sesh.cgroup, sesh.mean_charuco_fr_mar_xyz = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength, spoof_anipose_fail_bool=False)
+
+        sesh.cgroup, sesh.mean_charuco_fr_mar_xyz = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength)
+
+        ##this is supposed to cycle through tyhe videos with different windows to try to get Anipose to work. I can't get the dang thing working because something weird happens where a Thread will get spawned in one of the inner functions (maybe related to tqdm?) and that iteration will jump out of the try/except 
+        # try:
+        #     sesh.cgroup, sesh.mean_charuco_fr_mar_xyz = calibrate.CalibrateCaptureVolume(sesh,board, calVideoFrameLength)
+        #     # anipose_success = True
+        #     anipose_success = False
+        # except:
+        #     console.print_exception()
+        #     console.print('[bold red] - Anipose Calibration failed with user-provided (or default) `calVideoFrameLength` value! Trying again with other parts of the videos')
+        #     anipose_success = False
+
+        # if not anipose_success:
+        #     if calVideoFrameLength ==.25:
+        #         cal_video_frame_range = [round(sesh.numFrames*.25), round(sesh.numFrames*.5)]
+        #     else:
+        #         cal_video_frame_range = [0, round(sesh.numFrames*.25)]
[…]
