### 2021-11-02 3a99d02
Lots of aesthetic and 'usability' changes to print statements and labels
