+++ freemocap/fmc_blender.py
@@ -1,33 +1,60 @@
 import numpy as np, bpy
 from mathutils import Vector
 import math
-from pathlib import Path
 import sys
+from pathlib import Path
 
-#the path of the npy file 
+# get arguments
 argv = sys.argv
-argv = argv[argv.index("--") + 1:]  
+argv = argv[argv.index("--") + 1:] 
+ 
+# get npy session data
 input_npy = argv[0]
 
+# get session path
+sesh_path = argv[1]
+
+#the path of the npy build data file 
+build_data = Path(__file__).parent.resolve() / "build_data.npy"
+
 #3D array holding [[[x, y, z], [x, y, z]], [[x, y, z], [x, y, z]]] 
-arr = np.load(input_npy) 
+arr = np.load(input_npy)
 
-#the first frame
-markers_list = arr[0]
+# Saved clean frame where all points are visible and not NANs 
+markers_list = np.load(build_data)
 
-#a list containing icosphere objects
-order_of_icospheres = []
+#a list containing empty objects
+order_of_markers = []
 
 #frame rate 
 bpy.context.scene.render.fps = 30
+
+#dictionary corresponding to marker numbers and body parts
+body_dict = {
+    0:"head1", 
+    11:"shoulder_L", 
+    12:"shoulder_R", 
+    13:"elbow_L", 
+    14:"elbow_R", 
+    23:"hip_L",
+    24:"hip_R", 
+    25:"knee_L", 
+    26:"knee_R",
+    27:"ankle_L",
+    28:"ankle_R",
+    29:"heel_L", 
+    30:"heel_R",
+    31:"toe_L",
+    32:"toe_R" }
 
 #-----------------------------------------------------------------------------------   
 
 # set project unit 
 bpy.context.scene.unit_settings.length_unit = 'METERS'
 #iterate through arr and create an empty object at that location for each element
-for col in markers_list:
+for index, col in enumerate(markers_list):
     # parse string float value into floats, create Vector, set position to Vector
+
     if math.isnan(col[0]):
         col[0] = 0.0
     if math.isnan(col[1]):
@@ -35,15 +62,256 @@ for col in markers_list:
     if math.isnan(col[2]):
         col[2] = 0.0
     coord = Vector(((float(col[0])*0.001), (float(col[2])*0.001), (float(col[1]))* -0.001))
-    #add icosphere
-    bpy.ops.mesh.primitive_ico_sphere_add(subdivisions=1, radius=0.01, enter_editmode=False, location=coord)
-    sphere = bpy.context.active_object  
-    sphere.name = "Sphere"
-    order_of_icospheres.append(sphere)
+        
+    #empties
+    bpy.ops.object.add(type='EMPTY', location=coord)  
+    mt = bpy.context.active_object  
+    mt.name = "mt_" + str(index)
+    if index in body_dict.keys():
+        mt.name += "_" + str(body_dict[index])
+    order_of_markers.append(mt)
     #link empty to scene
-    bpy.context.scene.collection.objects.link( sphere )
+    bpy.context.scene.collection.objects.link( mt )
     #set location 
-    sphere.location = coord
+    mt.location = coord
+    #set the display size of the empty
+    mt.empty_display_size = 0.02
+    
+#--------------------------------------------------------------
+#Virtual Markers!
+
+# marker relationships:
+
+# - "weight": the weighted average of multiple markers. the virtual_markers[x] contains the 
+#list of markers that affect this virtual one. the weights[x] contains their corresponding weights in order.
+
+bpy.ops.object.mode_set(mode='OBJECT', toggle=False)
+
+#Create virtual marker where parameters are the name of the marker, the markers that affect its position, and each of their weights.
+def create_marker_weight(name, markers, weighted):
[…]
