+++ freemocap/fmc_runme.py
@@ -183,11 +183,11 @@ def RunMe(sessionID=None,
 
 
 
-    if useBlender == True:
-        #blenderExePath = Path('C:\Program Files\Blender Foundation\Blender 2.93')
-        #os.chdir(blenderExePath)
-        output = subprocess.run([blenderEXEpath, "--background", "--python", str(subprocessPath), "--", str(sesh.dataArrayPath/'mediaPipeSkel_3d.npy')], capture_output=True, text=True, check=True)
-        print(output)        
+    # if useBlender == True:
+    #     #blenderExePath = Path('C:\Program Files\Blender Foundation\Blender 2.93')
+    #     #os.chdir(blenderExePath)
+    #     output = subprocess.run([blenderEXEpath, "--background", "--python", str(subprocessPath), "--", str(sesh.dataArrayPath/'mediaPipeSkel_3d.npy')], capture_output=True, text=True, check=True)
+    #     print(output)        
 
     # %% Stage Five - Make Skreleton Animation
     if stage <= 5:
