### 2022-09-01 #41 @lwasser
hi there @jonmatthis ! this issue has been quiet for almost exactly a year so for now I am going to close it. We will be working out procedures that will allow the community to ask for help on package submissions - this likely will involve our [discourse forum](https://pyopensci.discourse.group/). If you have specific questions about python packaging I highly suggest that you ask the question there, and we will share via twitter to try to get some community engagement around supporting your efforts! In the future we will define a much more clear process to support efforts such as this so please also stay tuned for that and follow us here and on twitter (@pyopensci ) for updates and such. All the best.
